% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisAlignment-methods.R
\docType{methods}
\name{analysisAlignmentClient,OpencgaR-method}
\alias{analysisAlignmentClient,OpencgaR-method}
\title{analysisAlignmentClient methods}
\usage{
\S4method{analysisAlignmentClient}{OpencgaR}(OpencgaR, action, params = NULL,
  ...)
}
\arguments{
\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{action}{action to be performed on the variableSet(s)}

\item{params}{list containing additional query or body params}
}
\description{
This function implements the OpenCGA calls for managing Alignments
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
