% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisClinicalInterpretation-methods.R,
%   R/AnalysisVariant-methods.R
\docType{methods}
\name{analysisVariantClient,OpencgaR-method}
\alias{analysisVariantClient,OpencgaR-method}
\alias{analysisVariantClient,OpencgaR-method}
\title{analysisClinicalInterpretationClient methods}
\usage{
\S4method{analysisVariantClient}{OpencgaR}(OpencgaR, action, params = NULL,
  ...)

\S4method{analysisVariantClient}{OpencgaR}(OpencgaR, action, params = NULL,
  ...)
}
\arguments{
\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{action}{action to be performed on the variableSet(s)}

\item{params}{list containing additional query or body params}

\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{action}{action to be performed on the variableSet(s)}

\item{params}{list containing additional query or body params}
}
\description{
This function implements the OpenCGA calls for managing Cinical Interpretations

This function implements the OpenCGA calls for managing Variants
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}

\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
