% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisVariant-methods.R,
%   R/Clinical-methods.R, R/Cohort-methods.R, R/Sample-methods.R,
%   R/VariableSet-methods.R
\docType{data}
\name{category}
\alias{category}
\alias{category}
\alias{category}
\alias{category}
\alias{category}
\title{analysisVariantClient methods}
\format{An object of class \code{character} of length 1.}
\usage{
category

category

category

category

category
}
\arguments{
\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{action}{action to be performed on the variableSet(s)}

\item{params}{list containing additional query or body params}

\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{clinicalAnalysis}{a character string or vector containing clinical analysis IDs}

\item{action}{action to be performed on the sample(s)}

\item{params}{list containing additional query or body params}

\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{cohort}{a character string containing the cohort ID or name}

\item{annotationsetName}{a character string with the annotationset name. Only 
necessary when updating and deleting using the sampleAnnotationClient}

\item{memberId}{a character or a vector contaning user or group ids to 
work on. Mandatory when using sampleAclClient}

\item{action}{action to be performed on the sample(s)}

\item{params}{list containing additional query or body params}

\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{sample}{a character string or a vector containing sample IDs or names}

\item{annotationsetName}{a character string with the annotationset name. Only 
necessary when updating and deleting using the sampleAnnotationClient}

\item{memberId}{a character or a vector contaning user or group ids to 
work on. Mandatory when using sampleAclClient}

\item{action}{action to be performed on the sample(s)}

\item{params}{list containing additional query or body params}

\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{variableSet}{a character string or vector containing clinical analysis IDs}

\item{action}{action to be performed on the variableSet(s)}

\item{params}{list containing additional query or body params}
}
\description{
This function implements the OpenCGA calls for managing VariableSets

This function implements the OpenCGA calls for managing Clinical Analyses

This function implements the OpenCGA calls for managing Cohorts

This function implements the OpenCGA calls for managing Samples

This function implements the OpenCGA calls for managing VariableSets
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}

\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}

\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}

\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}

\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
\keyword{datasets}
