% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaClasses.R
\name{cgaVariantParam}
\alias{cgaVariantParam}
\title{A class to store Opencga parameters for filtering queries}
\usage{
cgaVariantParam(region = character(), chromosome = character(),
  type = character(), gene = character(), maf = character(),
  mgf = character(), genotype = character(), polyphen = character(),
  sift = character(), conservation = character(), reference = character(),
  alternate = character(), so = character(), biotype = character(),
  limit = character(), files = character(), returnedStudies = character(),
  returnedSamples = character())
}
\arguments{
\item{chromosome}{the name of the chromosome to be returned}
}
\description{
This class is used to store Opencga parameters for filtering queries
}
