% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort-methods.R
\docType{methods}
\name{cohortClient,OpencgaR-method}
\alias{cohortClient,OpencgaR-method}
\title{CohortClient methods}
\usage{
\S4method{cohortClient}{OpencgaR}(OpencgaR, cohort, action, params = NULL)
}
\arguments{
\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{cohort}{a character string containing the cohort ID or name}

\item{action}{action to be performed on the sample(s)}

\item{params}{list containing additional query or body params}

\item{annotationsetName}{a character string with the annotationset name. Only 
necessary when updating and deleting using the sampleAnnotationClient}

\item{memberId}{a character or a vector contaning user or group ids to 
work on. Mandatory when using sampleAclClient}
}
\description{
This function implements the OpenCGA calls for managing Cohorts
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
