% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVDB-methods.R
\name{cvdbClient,OpencgaR-method}
\alias{cvdbClient,OpencgaR-method}
\title{CVDBClient methods}
\usage{
\S4method{cvdbClient}{OpencgaR}(OpencgaR, variantId, endpointName, params = NULL, ...)
}
\arguments{
\item{variantId}{Variant ID (or comma separated list of variant IDs).}

\item{jobId}{Job ID. It must be a unique string within the study. An ID will be autogenerated automatically if not provided.}

\item{jobDescription}{Job description.}

\item{jobDependsOn}{Comma separated list of existing job IDs the job will depend on.}

\item{jobTags}{Job tags.}

\item{jobScheduledStartTime}{Time when the job is scheduled to start.}

\item{jobPriority}{Priority of the job.}

\item{jobDryRun}{Flag indicating that the job will be executed in dry-run mode. In this mode, OpenCGA will validate that all parameters and prerequisites are correctly set for successful execution, but the job will not actually run.}

\item{data}{Parameters: Index clinical analyses of a OpenCGA project, a study or a list of clinical analyses into CVDB.}

\item{field}{List of facet fields separated by semicolons, e.g.: type;biotypes. For nested faceted fields use >>, e.g.: type>>biotypes. Accepted values: caId, ciId, variantId, studyId, primary, discussionAuthor, discussionDate,discussionText, confidenceValue, confidenceAuthor, confidenceDate, tags, status, chromosome, start, end, xrefs, type, release, studies, phastCons, phylop, gerp, caddRaw, caddScaled, sift, siftDesc, polyphen, polyphenDesc, genes, biotypes, soAcc, clinicalSig.}

\item{study}{Study ID (or list of study IDs separated by commas).}

\item{include}{Fields included in the response, whole JSON path must be provided.}

\item{exclude}{Fields excluded in the response, whole JSON path must be provided.}

\item{limit}{Number of results to be returned.}

\item{skip}{Number of results to skip.}

\item{caId}{Clinical analysis ID (or list of IDs separated by commas).}

\item{caDescription}{Clinical analysis description (word or list of words contained in the text, if the words are separated by a comma an OR will be applied; if the words are separated by a semicolon, an AND will be applied).}

\item{caType}{Clinical analysis type (or list of types separated by commas).}

\item{caDisorderId}{Clinical analysis disorder ID (or list of IDs separated by commas).}

\item{caFilename}{Clinical analysis filename (or list of filenames separated by commas).}

\item{caProbandId}{Clinical analysis proband ID (or list of IDs separated by commas).}

\item{caProbandDisorderId}{Clinical analysis proband disorder ID (or list of disorder IDs separated by commas).}

\item{caProbandPhenotypeName}{Clinical analysis proband phenotype name (or list of phenotype names separated by commas).}

\item{caFamilyId}{Clinical analysis family ID (or list of IDs separated by commas).}

\item{caFamilyPhenotypeName}{Clinical analysis family phenotype names (or list of names separated by commas).}

\item{caFamilyMemberId}{Clinical analysis family member ID (or list of IDs separated by commas).}

\item{caReport}{Clinical analysis report text (word or list of words contained in the text, if the words are separated by a comma an OR will be applied; if the words are separated by a semicolon, an AND will be applied).}

\item{caStatus}{Clinical analysis status (or list of status separated by commas).}

\item{caLocked}{Clinical analysis locked (true or false).}

\item{ciId}{Clinical interpretation ID (or list of IDs separated by commas).}

\item{ciPrimary}{Clinical interpretation primary (true or false).}

\item{ciDescription}{Clinical interpretation description (word or list of words contained in the text, if the words are separated by a comma an OR will be applied; if the words are separated by a semicolon, an AND will be applied).}

\item{ciPanelId}{Clinical interpretation panel ID or name (or list of IDs or names separated by commas).}

\item{ciAnalystId}{Clinical interpretation analyst ID (or list of IDs separated by commas).}

\item{ciAnalystName}{Clinical interpretation analyst name (or list of names separated by commas).}

\item{ciAnalystEmail}{Clinical interpretation analyst e-mail (or list of e-mails separated by commas).}

\item{ciAnalystAssignedBy}{Clinical interpretation analyst assignee name (or list of names separated by commas).}

\item{ciAnalystDate}{Clinical interpretation analyst date (or list of dates separated by commas), with format YYYYMMDDhhmmss, e.g.: 20231026120345; range is available start_date-end_date, e.g.: 20231001000000-20231101000000.}

\item{ciMethodName}{Clinical interpretation method name (or list of names separated by commas).}

\item{ciMethodVersion}{Clinical interpretation method version (or list of versions separated by commas).}

\item{ciMethodCommit}{Clinical interpretation method commit (or list of commits separated by commas).}

\item{ciMethodDependencies}{Clinical interpretation method dependencies (word or list of words contained in the text, if the words are separated by a comma an OR will be applied; if the words are separated by a semicolon, an AND will be applied).}

\item{ciComments}{Clinical interpretation comment text (word or list of words contained in the text, if the words are separated by a comma an OR will be applied; if the words are separated by a semicolon, an AND will be applied).}

\item{ciLocked}{Clinical interpretation locked (true or false).}

\item{ciStatusId}{Clinical interpretation status ID (or list of IDs separated by commas). Valid values: NOT_STARTED, ACTIVE, DONE, CLOSED.}

\item{ciStatusName}{Clinical interpretation status name (or list of names separated by commas).}

\item{ciStatusDescription}{Clinical interpretation status description (word or list of words contained in the text, if the words are separated by a comma an OR will be applied; if the words are separated by a semicolon, an AND will be applied).}

\item{ciStatusDate}{Clinical interpretation status date (or list of dates separated by commas), with format YYYYMMDDhhmmss, e.g.: 20231026120345; range is available start_date-end_date, e.g.: 20231001000000-20231101000000.}

\item{ciCreationDate}{Clinical interpretation creation date (or list of dates separated by commas), with format YYYYMMDDhhmmss, e.g.: 20231026120345; range is available start_date-end_date, e.g.: 20231001000000-20231101000000.}

\item{ciModificationDate}{Clinical interpretation modification date (or list of dates separated by commas), with format YYYYMMDDhhmmss, e.g.: 20231026120345; range is available start_date-end_date, e.g.: 20231001000000-20231101000000.}

\item{ciVersion}{Clinical interpretation version number (or list of versions separated by commas).}

\item{cvId}{Clinical variant ID (or list of IDs separated by commas).}

\item{cvVariantId}{Variant ID (or list of IDs separated by commas).}

\item{cvPrimaryFinding}{Clinical variant is a primary finding (true or false).}

\item{cvPrimaryInterpretation}{Clinical variant belongs to the primary interpretation (true or false).}

\item{cvComments}{Clinical variant comment text (word or list of words contained in the text, if the words are separated by a comma an OR will be applied; if the words are separated by a semicolon, an AND will be applied).}

\item{cvDiscussionAuthor}{Clinical variant discussion author (or list of authors separated by commas).}

\item{cvDiscussionDate}{Clinical variant discussion date (or list of dates separated by commas), with format YYYYMMDDhhmmss, e.g.: 20231026120345; range is available start_date-end_date, e.g.: 20231001000000-20231101000000.}

\item{cvDiscussionText}{Clinical variant discussion text (word or list of words contained in the text, if the words are separated by a comma an OR will be applied; if the words are separated by a semicolon, an AND will be applied).}

\item{cvConfidenceValue}{Clinical variant confidence value (or list of values separated by commas).}

\item{cvConfidenceAuthor}{Clinical variant confidence author (or list of authors separated by commas).}

\item{cvConfidenceDate}{Clinical variant confidence date (or list of dates separated by commas), with format YYYYMMDDhhmmss, e.g.: 20231026120345; range is available start_date-end_date, e.g.: 20231001000000-20231101000000.}

\item{cvTag}{Clinical variant tag (or list of tags separated by commas).}

\item{cvStatus}{Clinical variant status (or list of status separated by commas).}

\item{cvRegion}{Variant region (or list of regions, these can be just a single chromosome name or regions in the format chr:start-end, e.g.: 2,3:100000-200000).}

\item{cvBiotype}{Variant biotype, e.g. protein_coding (or list of biotypes separated by commas).}

\item{cvCt}{Variant SO consequence type (or list of SOs separated by commas), e.g. missense_variant,stop_lost or SO:0001583,SO:0001578. Accepts aliases 'loss_of_function' and 'protein_altering'.}

\item{cvTranscriptFlag}{Variant transcript flag (or list of flags separated by commas), e.g. canonical, CCDS, basic, LRG, MANE Select, MANE Plus Clinical, EGLH_HaemOnc, TSO500.}

\item{cvGene}{Variant gene (or list genes separated by commas), most gene IDs are accepted (HGNC, Ensembl gene, ...).}

\item{cvXref}{Variant external reference (or list of references separated by commas), these can be genes, proteins or variants. Accepted IDs include HGNC, Ensembl genes, dbSNP, ClinVar, HPO, Cosmic, ...}

\item{cvAnnotRoleInCancerGenes}{Variant rol in cancer genes (or list of roles separated by commas).}

\item{cvType}{Variant type or list of types, accepted values are SNV, MNV, INDEL, SV, COPY_NUMBER, COPY_NUMBER_LOSS, COPY_NUMBER_GAIN, INSERTION, DELETION, DUPLICATION, TANDEM_DUPLICATION, BREAKEND, e.g. SNV,INDEL.}

\item{cvProteinSubstitution}{Variant protein substitution score (or list of scores separated by commas), include SIFT and PolyPhen. You can query using the score {protein_score}[<|>|<=|>=]{number} or the description {protein_score}[~=|=]{description} e.g. polyphen>0.1,sift=tolerant.}

\item{cvConservation}{Variant conservation score (or list of scores separated by commas) with the format {conservation_score}[<|>|<=|>=]{number} e.g. phastCons>0.5,phylop<0.1,gerp>0.1.}

\item{cvFunctionalScore}{Variant functional score (or list of scores separated by commas) with the format {functional_score}[<|>|<=|>=]{number} e.g. cadd_scaled>5.2 , cadd_raw<=0.3.}

\item{cvPopulationFrequencyAlt}{Variant alternate population frequency (or list of frequencies separated by commas), with the format {study}:{population}[<|>|<=|>=]{number}. e.g. 1000G:ALL<0.01.}

\item{cvPopulationFrequencyMaf}{Variant population minor allele frequency (or list of frequencies separated by commas), with the format {study}:{population}[<|>|<=|>=]{number}. e.g. 1000G:ALL<0.01.}

\item{cvPopulationFrequencyRef}{Variant reference population frequency (or list of frequences separated by commas), with the format {study}:{population}[<|>|<=|>=]{number}. e.g. 1000G:ALL<0.01.}

\item{cvCohortStatsAlt}{Variant alternate allele frequency (or list of frequencies separated by commas), with the format [{study:}]{cohort}[<|>|<=|>=]{number}. e.g. ALL<=0.4.}

\item{cvCohortStatsMaf}{Variant minor allele frequency (or list of frequencies separated by commas), with the format [{study:}]{cohort}[<|>|<=|>=]{number}. e.g. ALL<=0.4.}

\item{cvCohortStatsRef}{Variant reference allele frequency (or list of frequencies separated by commas), with the foramt [{study:}]{cohort}[<|>|<=|>=]{number}. e.g. ALL<=0.4.}

\item{cvCohortStatsPass}{Variant filter PASS frequency (or list of frequencies separated by commas), with the format [{study:}]{cohort}[<|>|<=|>=]{number}. e.g. ALL>0.8.}

\item{cvScore}{Variant score (or list of scores separated by commas), with the format: [{study:}]{score}[<|>|<=|>=]{number}.}

\item{cvAnnotGoGenes}{Variant gene GO (or list of GOs separated by commas).}

\item{cvAnnotExpressionGenes}{Variant gene expression (or list of expressions separated by commas).}

\item{cvGeneTraitId}{Variant gene trait association ID (or list of trait IDs separated by commas), e.g. "umls:C0007222" , "OMIM:269600".}

\item{cvTrait}{Variant Trait (or list of traits separated by commas), based on ClinVar, HPO, COSMIC, i.e.: IDs, histologies, descriptions,...}

\item{cvProteinKeyword}{Uniprot protein variant annotation keyword (or list of keywords separated by commas).}

\item{cveVariantId}{Variant ID (or list of IDs separated by commas).}

\item{cvePrimaryFinding}{Clinical variant evidence belongs to a primary finding (true or false).}

\item{cvePrimaryInterpretation}{Clinical variant evidence belongs to the primary interpretation (true or false).}

\item{cvePhenotypeName}{Clinical variant evidence phenotype name (or names separated by commas).}

\item{cveGeneName}{Clinical variant evidence gene name (or names separated by commas).}

\item{cveTranscriptId}{Clinical variant evidence transcript ID (or IDs separated by commas).}

\item{cveSoTermName}{Clinical variant evidence sequence ontology name (or names separated by commas).}

\item{cveXrefId}{Clinical variant evidence Xref ID (or IDs separated by commas).}

\item{cvePanelId}{Clinical variant evidence panel ID (or IDs separated by commas).}

\item{cveMoi}{Clinical variant evidence mode of inheritance (or list of modes of inheritance separated by commas), valid values: AUTOSOMAL_DOMINANT, AUTOSOMAL_RECESSIVE, X_LINKED_DOMINANT, X_LINKED_RECESSIVE, Y_LINKED, MITOCHONDRIAL, DE_NOVO, MENDELIAN_ERROR, COMPOUND_HETEROZYGOUS, UNKNOWN.}

\item{cvePenetrance}{Clinical variant evidence penetrance (or list of penetrance values separated by commas), valid values: COMPLETE, INCOMPLETE, UNKNOWN.}

\item{cveAcmg}{Clinical variant evidence ACMG (or ACGMs separated by commas).}

\item{cveTier}{Clinical variant evidence tier (or list of tier values separated by commas).}

\item{cveClinicalSignificance}{Clinical variant evidence clinical significance (or list of clinical  significances separated by commas).}

\item{cveDrugResponse}{Clinical variant evidence drug response (or list of drug responses separated by commas).}

\item{cveTraitAssociation}{Clinical variant evidence trait association (or list of traits separated by commas).}

\item{cveFunctionalEffect}{Clinical variant evidence functional effect (or list of functional effects separated by commas).}

\item{cveTumorigenesis}{Clinical variant evidence tumorigenesis (or list of tumorigenesis values separated by commas).}

\item{cveOtherClassification}{Clinical variant evidence other-classification (or list of other  classification values separated by commas).}

\item{cveRoleInCancer}{Clinical variant evidence role in cancer (or roles in cancer separated by commas).}

\item{cveReviewAcmg}{Clinical variant evidence review ACMG (or ACGMs separated by commas).}

\item{cveReviewTier}{Clinical variant evidence review tier (or list of tier values separated by commas).}

\item{cveReviewClinicalSignificance}{Clinical variant evidence review clinical significance (or list of clinical significances separated by commas).}

\item{cveReviewText}{Clinical variant evidence review text (word or list of words contained in the text, if the words are separated by a comma an OR will be applied; if the words are separated by a semicolon, an AND will be applied).}

\item{project}{Project ID(or command separated list of project IDs).}
}
\description{
This function implements the OpenCGA calls for managing Analysis - CVDB.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   aggregateAnalysis \tab /{apiVersion}/analysis/cvdb/analysis/aggregate \tab project, study, caId, caDescription, caType, caDisorderId, caFilename, caProbandId, caProbandDisorderId, caProbandPhenotypeName, caFamilyId, caFamilyPhenotypeName, caFamilyMemberId, caReport, caStatus, caLocked, ciId, ciPrimary, ciDescription, ciPanelId, ciAnalystId, ciAnalystName, ciAnalystEmail, ciAnalystAssignedBy, ciAnalystDate, ciMethodName, ciMethodVersion, ciMethodCommit, ciMethodDependencies, ciComments, ciLocked, ciStatusId, ciStatusName, ciStatusDescription, ciStatusDate, ciCreationDate, ciModificationDate, ciVersion, cvId, cvVariantId, cvPrimaryFinding, cvPrimaryInterpretation, cvComments, cvDiscussionAuthor, cvDiscussionDate, cvDiscussionText, cvConfidenceValue, cvConfidenceAuthor, cvConfidenceDate, cvTag, cvStatus, cvRegion, cvBiotype, cvCt, cvTranscriptFlag, cvGene, cvXref, cvAnnotRoleInCancerGenes, cvType, cvProteinSubstitution, cvConservation, cvFunctionalScore, cvPopulationFrequencyAlt, cvPopulationFrequencyMaf, cvPopulationFrequencyRef, cvCohortStatsAlt, cvCohortStatsMaf, cvCohortStatsRef, cvCohortStatsPass, cvScore, cvAnnotGoGenes, cvAnnotExpressionGenes, cvGeneTraitId, cvTrait, cvProteinKeyword, cveVariantId, cvePrimaryFinding, cvePrimaryInterpretation, cvePhenotypeName, cveGeneName, cveTranscriptId, cveSoTermName, cveXrefId, cvePanelId, cveMoi, cvePenetrance, cveAcmg, cveTier, cveClinicalSignificance, cveDrugResponse, cveTraitAssociation, cveFunctionalEffect, cveTumorigenesis, cveOtherClassification, cveRoleInCancer, cveReviewAcmg, cveReviewTier, cveReviewClinicalSignificance, cveReviewText, field \cr
   queryAnalysis \tab /{apiVersion}/analysis/cvdb/analysis/query \tab project, study, include, exclude, limit, skip, caId, caDescription, caType, caDisorderId, caFilename, caProbandId, caProbandDisorderId, caProbandPhenotypeName, caFamilyId, caFamilyPhenotypeName, caFamilyMemberId, caReport, caStatus, caLocked, ciId, ciPrimary, ciDescription, ciPanelId, ciAnalystId, ciAnalystName, ciAnalystEmail, ciAnalystAssignedBy, ciAnalystDate, ciMethodName, ciMethodVersion, ciMethodCommit, ciMethodDependencies, ciComments, ciLocked, ciStatusId, ciStatusName, ciStatusDescription, ciStatusDate, ciCreationDate, ciModificationDate, ciVersion, cvId, cvVariantId, cvPrimaryFinding, cvPrimaryInterpretation, cvComments, cvDiscussionAuthor, cvDiscussionDate, cvDiscussionText, cvConfidenceValue, cvConfidenceAuthor, cvConfidenceDate, cvTag, cvStatus, cvRegion, cvBiotype, cvCt, cvTranscriptFlag, cvGene, cvXref, cvAnnotRoleInCancerGenes, cvType, cvProteinSubstitution, cvConservation, cvFunctionalScore, cvPopulationFrequencyAlt, cvPopulationFrequencyMaf, cvPopulationFrequencyRef, cvCohortStatsAlt, cvCohortStatsMaf, cvCohortStatsRef, cvCohortStatsPass, cvScore, cvAnnotGoGenes, cvAnnotExpressionGenes, cvGeneTraitId, cvTrait, cvProteinKeyword, cveVariantId, cvePrimaryFinding, cvePrimaryInterpretation, cvePhenotypeName, cveGeneName, cveTranscriptId, cveSoTermName, cveXrefId, cvePanelId, cveMoi, cvePenetrance, cveAcmg, cveTier, cveClinicalSignificance, cveDrugResponse, cveTraitAssociation, cveFunctionalEffect, cveTumorigenesis, cveOtherClassification, cveRoleInCancer, cveReviewAcmg, cveReviewTier, cveReviewClinicalSignificance, cveReviewText \cr
   aggregateEvidence \tab /{apiVersion}/analysis/cvdb/evidence/aggregate \tab project, study, caId, caDescription, caType, caDisorderId, caFilename, caProbandId, caProbandDisorderId, caProbandPhenotypeName, caFamilyId, caFamilyPhenotypeName, caFamilyMemberId, caReport, caStatus, caLocked, ciId, ciPrimary, ciDescription, ciPanelId, ciAnalystId, ciAnalystName, ciAnalystEmail, ciAnalystAssignedBy, ciAnalystDate, ciMethodName, ciMethodVersion, ciMethodCommit, ciMethodDependencies, ciComments, ciLocked, ciStatusId, ciStatusName, ciStatusDescription, ciStatusDate, ciCreationDate, ciModificationDate, ciVersion, cvId, cvVariantId, cvPrimaryFinding, cvPrimaryInterpretation, cvComments, cvDiscussionAuthor, cvDiscussionDate, cvDiscussionText, cvConfidenceValue, cvConfidenceAuthor, cvConfidenceDate, cvTag, cvStatus, cvRegion, cvBiotype, cvCt, cvTranscriptFlag, cvGene, cvXref, cvAnnotRoleInCancerGenes, cvType, cvProteinSubstitution, cvConservation, cvFunctionalScore, cvPopulationFrequencyAlt, cvPopulationFrequencyMaf, cvPopulationFrequencyRef, cvCohortStatsAlt, cvCohortStatsMaf, cvCohortStatsRef, cvCohortStatsPass, cvScore, cvAnnotGoGenes, cvAnnotExpressionGenes, cvGeneTraitId, cvTrait, cvProteinKeyword, cveVariantId, cvePrimaryFinding, cvePrimaryInterpretation, cvePhenotypeName, cveGeneName, cveTranscriptId, cveSoTermName, cveXrefId, cvePanelId, cveMoi, cvePenetrance, cveAcmg, cveTier, cveClinicalSignificance, cveDrugResponse, cveTraitAssociation, cveFunctionalEffect, cveTumorigenesis, cveOtherClassification, cveRoleInCancer, cveReviewAcmg, cveReviewTier, cveReviewClinicalSignificance, cveReviewText, field \cr
   queryEvidence \tab /{apiVersion}/analysis/cvdb/evidence/query \tab project, study, include, exclude, limit, skip, caId, caDescription, caType, caDisorderId, caFilename, caProbandId, caProbandDisorderId, caProbandPhenotypeName, caFamilyId, caFamilyPhenotypeName, caFamilyMemberId, caReport, caStatus, caLocked, ciId, ciPrimary, ciDescription, ciPanelId, ciAnalystId, ciAnalystName, ciAnalystEmail, ciAnalystAssignedBy, ciAnalystDate, ciMethodName, ciMethodVersion, ciMethodCommit, ciMethodDependencies, ciComments, ciLocked, ciStatusId, ciStatusName, ciStatusDescription, ciStatusDate, ciCreationDate, ciModificationDate, ciVersion, cvId, cvVariantId, cvPrimaryFinding, cvPrimaryInterpretation, cvComments, cvDiscussionAuthor, cvDiscussionDate, cvDiscussionText, cvConfidenceValue, cvConfidenceAuthor, cvConfidenceDate, cvTag, cvStatus, cvRegion, cvBiotype, cvCt, cvTranscriptFlag, cvGene, cvXref, cvAnnotRoleInCancerGenes, cvType, cvProteinSubstitution, cvConservation, cvFunctionalScore, cvPopulationFrequencyAlt, cvPopulationFrequencyMaf, cvPopulationFrequencyRef, cvCohortStatsAlt, cvCohortStatsMaf, cvCohortStatsRef, cvCohortStatsPass, cvScore, cvAnnotGoGenes, cvAnnotExpressionGenes, cvGeneTraitId, cvTrait, cvProteinKeyword, cveVariantId, cvePrimaryFinding, cvePrimaryInterpretation, cvePhenotypeName, cveGeneName, cveTranscriptId, cveSoTermName, cveXrefId, cvePanelId, cveMoi, cvePenetrance, cveAcmg, cveTier, cveClinicalSignificance, cveDrugResponse, cveTraitAssociation, cveFunctionalEffect, cveTumorigenesis, cveOtherClassification, cveRoleInCancer, cveReviewAcmg, cveReviewTier, cveReviewClinicalSignificance, cveReviewText \cr
   runIndex \tab /{apiVersion}/analysis/cvdb/index/run \tab study, jobId, jobDescription, jobDependsOn, jobTags, jobScheduledStartTime, jobPriority, jobDryRun, body\link{*} \cr
   aggregateInterpretation \tab /{apiVersion}/analysis/cvdb/interpretation/aggregate \tab project, study, caId, caDescription, caType, caDisorderId, caFilename, caProbandId, caProbandDisorderId, caProbandPhenotypeName, caFamilyId, caFamilyPhenotypeName, caFamilyMemberId, caReport, caStatus, caLocked, ciId, ciPrimary, ciDescription, ciPanelId, ciAnalystId, ciAnalystName, ciAnalystEmail, ciAnalystAssignedBy, ciAnalystDate, ciMethodName, ciMethodVersion, ciMethodCommit, ciMethodDependencies, ciComments, ciLocked, ciStatusId, ciStatusName, ciStatusDescription, ciStatusDate, ciCreationDate, ciModificationDate, ciVersion, cvId, cvVariantId, cvPrimaryFinding, cvPrimaryInterpretation, cvComments, cvDiscussionAuthor, cvDiscussionDate, cvDiscussionText, cvConfidenceValue, cvConfidenceAuthor, cvConfidenceDate, cvTag, cvStatus, cvRegion, cvBiotype, cvCt, cvTranscriptFlag, cvGene, cvXref, cvAnnotRoleInCancerGenes, cvType, cvProteinSubstitution, cvConservation, cvFunctionalScore, cvPopulationFrequencyAlt, cvPopulationFrequencyMaf, cvPopulationFrequencyRef, cvCohortStatsAlt, cvCohortStatsMaf, cvCohortStatsRef, cvCohortStatsPass, cvScore, cvAnnotGoGenes, cvAnnotExpressionGenes, cvGeneTraitId, cvTrait, cvProteinKeyword, cveVariantId, cvePrimaryFinding, cvePrimaryInterpretation, cvePhenotypeName, cveGeneName, cveTranscriptId, cveSoTermName, cveXrefId, cvePanelId, cveMoi, cvePenetrance, cveAcmg, cveTier, cveClinicalSignificance, cveDrugResponse, cveTraitAssociation, cveFunctionalEffect, cveTumorigenesis, cveOtherClassification, cveRoleInCancer, cveReviewAcmg, cveReviewTier, cveReviewClinicalSignificance, cveReviewText, field \cr
   queryInterpretation \tab /{apiVersion}/analysis/cvdb/interpretation/query \tab project, study, include, exclude, limit, skip, caId, caDescription, caType, caDisorderId, caFilename, caProbandId, caProbandDisorderId, caProbandPhenotypeName, caFamilyId, caFamilyPhenotypeName, caFamilyMemberId, caReport, caStatus, caLocked, ciId, ciPrimary, ciDescription, ciPanelId, ciAnalystId, ciAnalystName, ciAnalystEmail, ciAnalystAssignedBy, ciAnalystDate, ciMethodName, ciMethodVersion, ciMethodCommit, ciMethodDependencies, ciComments, ciLocked, ciStatusId, ciStatusName, ciStatusDescription, ciStatusDate, ciCreationDate, ciModificationDate, ciVersion, cvId, cvVariantId, cvPrimaryFinding, cvPrimaryInterpretation, cvComments, cvDiscussionAuthor, cvDiscussionDate, cvDiscussionText, cvConfidenceValue, cvConfidenceAuthor, cvConfidenceDate, cvTag, cvStatus, cvRegion, cvBiotype, cvCt, cvTranscriptFlag, cvGene, cvXref, cvAnnotRoleInCancerGenes, cvType, cvProteinSubstitution, cvConservation, cvFunctionalScore, cvPopulationFrequencyAlt, cvPopulationFrequencyMaf, cvPopulationFrequencyRef, cvCohortStatsAlt, cvCohortStatsMaf, cvCohortStatsRef, cvCohortStatsPass, cvScore, cvAnnotGoGenes, cvAnnotExpressionGenes, cvGeneTraitId, cvTrait, cvProteinKeyword, cveVariantId, cvePrimaryFinding, cvePrimaryInterpretation, cvePhenotypeName, cveGeneName, cveTranscriptId, cveSoTermName, cveXrefId, cvePanelId, cveMoi, cvePenetrance, cveAcmg, cveTier, cveClinicalSignificance, cveDrugResponse, cveTraitAssociation, cveFunctionalEffect, cveTumorigenesis, cveOtherClassification, cveRoleInCancer, cveReviewAcmg, cveReviewTier, cveReviewClinicalSignificance, cveReviewText \cr
   aggregateVariant \tab /{apiVersion}/analysis/cvdb/variant/aggregate \tab project, study, caId, caDescription, caType, caDisorderId, caFilename, caProbandId, caProbandDisorderId, caProbandPhenotypeName, caFamilyId, caFamilyPhenotypeName, caFamilyMemberId, caReport, caStatus, caLocked, ciId, ciPrimary, ciDescription, ciPanelId, ciAnalystId, ciAnalystName, ciAnalystEmail, ciAnalystAssignedBy, ciAnalystDate, ciMethodName, ciMethodVersion, ciMethodCommit, ciMethodDependencies, ciComments, ciLocked, ciStatusId, ciStatusName, ciStatusDescription, ciStatusDate, ciCreationDate, ciModificationDate, ciVersion, cvId, cvVariantId, cvPrimaryFinding, cvPrimaryInterpretation, cvComments, cvDiscussionAuthor, cvDiscussionDate, cvDiscussionText, cvConfidenceValue, cvConfidenceAuthor, cvConfidenceDate, cvTag, cvStatus, cvRegion, cvBiotype, cvCt, cvTranscriptFlag, cvGene, cvXref, cvAnnotRoleInCancerGenes, cvType, cvProteinSubstitution, cvConservation, cvFunctionalScore, cvPopulationFrequencyAlt, cvPopulationFrequencyMaf, cvPopulationFrequencyRef, cvCohortStatsAlt, cvCohortStatsMaf, cvCohortStatsRef, cvCohortStatsPass, cvScore, cvAnnotGoGenes, cvAnnotExpressionGenes, cvGeneTraitId, cvTrait, cvProteinKeyword, cveVariantId, cvePrimaryFinding, cvePrimaryInterpretation, cvePhenotypeName, cveGeneName, cveTranscriptId, cveSoTermName, cveXrefId, cvePanelId, cveMoi, cvePenetrance, cveAcmg, cveTier, cveClinicalSignificance, cveDrugResponse, cveTraitAssociation, cveFunctionalEffect, cveTumorigenesis, cveOtherClassification, cveRoleInCancer, cveReviewAcmg, cveReviewTier, cveReviewClinicalSignificance, cveReviewText, field \cr
   queryVariant \tab /{apiVersion}/analysis/cvdb/variant/query \tab project, study, include, exclude, limit, skip, caId, caDescription, caType, caDisorderId, caFilename, caProbandId, caProbandDisorderId, caProbandPhenotypeName, caFamilyId, caFamilyPhenotypeName, caFamilyMemberId, caReport, caStatus, caLocked, ciId, ciPrimary, ciDescription, ciPanelId, ciAnalystId, ciAnalystName, ciAnalystEmail, ciAnalystAssignedBy, ciAnalystDate, ciMethodName, ciMethodVersion, ciMethodCommit, ciMethodDependencies, ciComments, ciLocked, ciStatusId, ciStatusName, ciStatusDescription, ciStatusDate, ciCreationDate, ciModificationDate, ciVersion, cvId, cvVariantId, cvPrimaryFinding, cvPrimaryInterpretation, cvComments, cvDiscussionAuthor, cvDiscussionDate, cvDiscussionText, cvConfidenceValue, cvConfidenceAuthor, cvConfidenceDate, cvTag, cvStatus, cvRegion, cvBiotype, cvCt, cvTranscriptFlag, cvGene, cvXref, cvAnnotRoleInCancerGenes, cvType, cvProteinSubstitution, cvConservation, cvFunctionalScore, cvPopulationFrequencyAlt, cvPopulationFrequencyMaf, cvPopulationFrequencyRef, cvCohortStatsAlt, cvCohortStatsMaf, cvCohortStatsRef, cvCohortStatsPass, cvScore, cvAnnotGoGenes, cvAnnotExpressionGenes, cvGeneTraitId, cvTrait, cvProteinKeyword, cveVariantId, cvePrimaryFinding, cvePrimaryInterpretation, cvePhenotypeName, cveGeneName, cveTranscriptId, cveSoTermName, cveXrefId, cvePanelId, cveMoi, cvePenetrance, cveAcmg, cveTier, cveClinicalSignificance, cveDrugResponse, cveTraitAssociation, cveFunctionalEffect, cveTumorigenesis, cveOtherClassification, cveRoleInCancer, cveReviewAcmg, cveReviewTier, cveReviewClinicalSignificance, cveReviewText \cr
   statsVariant \tab /{apiVersion}/analysis/cvdb/variant/{variantId}/stats \tab variantId\link{*}, project \cr
}
}
\section{Endpoint /{apiVersion}/analysis/cvdb/analysis/aggregate}{

Calculate and fetch clinical analysis aggregation stats.
}

\section{Endpoint /{apiVersion}/analysis/cvdb/analysis/query}{

Filter and fetch clinical analyses from CVDB.
}

\section{Endpoint /{apiVersion}/analysis/cvdb/evidence/aggregate}{

Calculate and fetch clinical variant evidence aggregation stats.
}

\section{Endpoint /{apiVersion}/analysis/cvdb/evidence/query}{

Filter and fetch clinical variant evidences from CVDB.
}

\section{Endpoint /{apiVersion}/analysis/cvdb/index/run}{

Index clinical analyses of a OpenCGA project, a study or a list of clinical analyses into CVDB.
}

\section{Endpoint /{apiVersion}/analysis/cvdb/interpretation/aggregate}{

Calculate and fetch clinical interpretation aggregation stats.
}

\section{Endpoint /{apiVersion}/analysis/cvdb/interpretation/query}{

Filter and fetch clinical interpretations from CVDB.
}

\section{Endpoint /{apiVersion}/analysis/cvdb/variant/aggregate}{

Calculate and fetch clinical variant aggregation stats.
}

\section{Endpoint /{apiVersion}/analysis/cvdb/variant/query}{

Filter and fetch clinical variants from CVDB.
}

\section{Endpoint /{apiVersion}/analysis/cvdb/variant/{variantId}/stats}{

Get clinical variant summary from CVDB.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
