% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasePanel-methods.R
\docType{methods}
\name{diseasePanelClient,OpencgaR-method}
\alias{diseasePanelClient,OpencgaR-method}
\title{DiseasePanel methods}
\usage{
\S4method{diseasePanelClient}{OpencgaR}(OpencgaR, diseasePanel, action,
  params = NULL, ...)
}
\arguments{
\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{diseasePanel}{a string or vector containing the disease panel ID(s)}

\item{action}{action to be performed on the sample(s)}

\item{params}{list containing additional query or body params}
}
\description{
This function implements the OpenCGA calls for managing Disease Panels
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
