% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Family-methods.R
\docType{methods}
\name{familyClient,OpencgaR-method}
\alias{familyClient,OpencgaR-method}
\title{FamilyClient methods}
\usage{
\S4method{familyClient}{OpencgaR}(OpencgaR, family, action, params = NULL,
  ...)
}
\arguments{
\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{family}{a character string or a vector containing family ids}

\item{action}{action to be performed on the family or families}

\item{params}{list containing additional query or body params}

\item{annotationsetName}{a character string with the annotationset name. Only 
necessary when updating and deleting using the familyAnnotationClient}

\item{memberId}{a character or a vector contaning user or group ids to 
work on. Mandatory when using familyAclClient}
}
\description{
This function implements the OpenCGA calls for managing Jobs
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
