% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Federation-methods.R
\name{federationClient,OpencgaR-method}
\alias{federationClient,OpencgaR-method}
\title{FederationClient methods}
\usage{
\S4method{federationClient}{OpencgaR}(OpencgaR, id, endpointName, params = NULL, ...)
}
\arguments{
\item{id}{Federation server id.}

\item{study}{Study [\link{organization@}project:]study where study and project can be either the ID or UUID.}

\item{action}{Action to be performed: ADD access or REMOVE access. Allowed values: \link{'ADD', 'REMOVE'}}

\item{data}{JSON containing the Federation server parameters to be updated.}
}
\description{
This function implements the OpenCGA calls for managing Federations.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   connectClient \tab /{apiVersion}/federations/client/connect \tab body\link{*} \cr
   clientStudyUsersList \tab /{apiVersion}/federations/client/study/users/list \tab study \cr
   clientStudyUsersUpdate \tab /{apiVersion}/federations/client/study/users/update \tab study, action, body\link{*} \cr
   deleteClient \tab /{apiVersion}/federations/client/{id}/delete \tab id\link{*} \cr
   synchronizeClient \tab /{apiVersion}/federations/client/{id}/synchronize \tab id\link{*} \cr
   updateClient \tab /{apiVersion}/federations/client/{id}/update \tab id\link{*}, body\link{*} \cr
   createServer \tab /{apiVersion}/federations/server/create \tab body\link{*} \cr
   deleteServer \tab /{apiVersion}/federations/server/{id}/delete \tab id\link{*} \cr
   resetServer \tab /{apiVersion}/federations/server/{id}/reset \tab id\link{*} \cr
   updateServer \tab /{apiVersion}/federations/server/{id}/update \tab id\link{*}, body\link{*} \cr
}
}
\section{Endpoint /{apiVersion}/federations/client/connect}{

Connect to a shared XetaBase instance.
}

\section{Endpoint /{apiVersion}/federations/client/study/users/list}{

Show the list of users with access to the federated study.
}

\section{Endpoint /{apiVersion}/federations/client/study/users/update}{

Grant/Deny access to federated studies to users.
}

\section{Endpoint /{apiVersion}/federations/client/{id}/delete}{

Delete a federation client.
}

\section{Endpoint /{apiVersion}/federations/client/{id}/synchronize}{

Synchronize data from a known Federation server.
}

\section{Endpoint /{apiVersion}/federations/client/{id}/update}{

Update some fields from a Federation client.
}

\section{Endpoint /{apiVersion}/federations/server/create}{

Share a resource with another XetaBase instance.
}

\section{Endpoint /{apiVersion}/federations/server/{id}/delete}{

Delete a federation server.
}

\section{Endpoint /{apiVersion}/federations/server/{id}/reset}{

Reset the credentials of a federation server.
}

\section{Endpoint /{apiVersion}/federations/server/{id}/update}{

Update some fields from a Federation server.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
