% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File-methods.R
\docType{methods}
\name{fileClient,OpencgaR-method}
\alias{fileClient,OpencgaR-method}
\title{FileClient methods}
\usage{
\S4method{fileClient}{OpencgaR}(OpencgaR, f, action, params = NULL, ...)
}
\arguments{
\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{f}{a character string or a vector containing file ids, folder ids, fields 
or a path. When specifying paths they must be separated by ":" instead of "/".
Mandatory when using studyClient with all actions except "create" and "search" 
and all actions in studyGroupClient}

\item{action}{action to be performed on the files}

\item{params}{list containing additional query or body params}

\item{memberId}{a character or a vector contaning user or group ids to 
work on. Mandatory when using studyAclClient}
}
\description{
This function implements the OpenCGA calls for managing Files
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
