% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA4GH-methods.R
\name{ga4ghClient,OpencgaR-method}
\alias{ga4ghClient,OpencgaR-method}
\title{GA4GHClient methods}
\usage{
\S4method{ga4ghClient}{OpencgaR}(OpencgaR, study, file, endpointName, params = NULL, ...)
}
\arguments{
\item{study}{Study [\link{user@}project:]study where study and project can be either the ID or UUID.}

\item{file}{File id, name or path.}

\item{referenceName}{Reference sequence name (Example: 'chr1', '1' or 'chrX'.}

\item{start}{The start position of the range on the reference, 0-based, inclusive.}

\item{end}{The end position of the range on the reference, 0-based, exclusive.}

\item{referenceGenome}{Reference genome.}

\item{chrom}{Chromosome ID. Accepted values: 1-22, X, Y, MT. Note: For compatibility with conventions set by some of the existing beacons, an arbitrary prefix is accepted as well (e.g. chr1 is equivalent to chrom1 and 1).}

\item{pos}{Coordinate within a chromosome. Position is a number and is 0-based.}

\item{allele}{Any string of nucleotides A,C,T,G or D, I for deletion and insertion, respectively. Note: For compatibility with conventions set by some of the existing beacons, DEL and INS identifiers are also accepted.}

\item{ref}{Genome ID. If not specified, all the genomes supported by the given beacons are queried. Note: For compatibility with conventions set by some of the existing beacons, both GRC or HG notation are accepted, case insensitive.}

\item{beacon}{Beacon IDs. If specified, only beacons with the given IDs are queried. Responses from all the supported beacons are obtained otherwise. Format: \link{id1,id2}.}
}
\description{
This function implements the OpenCGA calls for managing GA4GH.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   searchReads \tab /{apiVersion}/ga4gh/reads/search \tab  \cr
   fetchReads \tab /{apiVersion}/ga4gh/reads/{study}/{file} \tab file\link{*}, study\link{*}, referenceName, start, end, referenceGenome \cr
   responses \tab /{apiVersion}/ga4gh/responses \tab chrom\link{*}, pos\link{*}, allele\link{*}, ref, beacon\link{*} \cr
   searchVariants \tab /{apiVersion}/ga4gh/variants/search \tab  \cr
}
}
\section{Endpoint /{apiVersion}/ga4gh/reads/search}{

Description.
}

\section{Endpoint /{apiVersion}/ga4gh/reads/{study}/{file}}{

Fetch alignment files using HTSget protocol.
}

\section{Endpoint /{apiVersion}/ga4gh/responses}{

Beacon webservice.
}

\section{Endpoint /{apiVersion}/ga4gh/variants/search}{

Description.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
