% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestResponse-methods.R
\name{getApiVersion}
\alias{getApiVersion}
\title{RestResponse methods}
\usage{
getApiVersion(restResponse)
}
\description{
The following methods implement the functions to extract 
information from the RestResponse object
}
\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the 
RESTful API documentation \url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
Getters from RestResponse
}
