% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaFunctions.R
\name{getOpencgaDocs}
\alias{getOpencgaDocs}
\title{A method to fetch documentation about Opencga Methods}
\usage{
getOpencgaDocs(category, action, requiredOnly = FALSE)
}
\arguments{
\item{category}{the Opencga category being queried}

\item{action}{action intended on that category}
}
\value{
A dataframe
}
\description{
This method allow users to lookup required 
to query Opencga Data
}
