% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual-methods.R
\docType{methods}
\name{individualClient,OpencgaR-method}
\alias{individualClient,OpencgaR-method}
\title{IndividualClient methods}
\usage{
\S4method{individualClient}{OpencgaR}(OpencgaR, individual, action,
  params = NULL, ...)
}
\arguments{
\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{individual}{a character string or a vector containing individual ids}

\item{action}{action to be performed on the individual(s)}

\item{params}{list containing additional query or body params}

\item{annotationsetName}{a character string with the annotationset name. Only 
necessary when updating and deleting using the individualAnnotationClient}

\item{memberId}{a character or a vector contaning user or group ids to 
work on. Mandatory when using individualAclClient}
}
\description{
This function implements the OpenCGA calls for managing Individuals
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
