% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaR-methods.R
\name{initOpencgaR}
\alias{initOpencgaR}
\title{initOpencgaR}
\usage{
initOpencgaR(host = NULL, version = "v1", user = "",
  opencgaConfig = NULL)
}
\arguments{
\item{host}{a character specifying the host url, e.g.
"http://bioinfo.hpc.cam.ac.uk/hgva"}

\item{version}{a character specifying the API version, e.g. "v1"}
}
\value{
An object of class OpencgaR
}
\description{
This function inicializes the OpencgaR object with the 
necessary details to generate a connection.
}
\details{
OpencgaR init function


This class initializes the OpencgaR object. It holds the default configuration 
required by OpencgaR methods to connect to OpenCGA web services. By defult, 
it is configured to query HGVA (http://hgva.opencb.org/).
}
\examples{
   ocga <- initOpencgaR()
   print(ocga)
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} 
and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
