% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job-methods.R
\docType{methods}
\name{jobClient,OpencgaR-method}
\alias{jobClient,OpencgaR-method}
\title{JobClient methods}
\usage{
\S4method{jobClient}{OpencgaR}(OpencgaR, jobId, action, params = NULL, ...)
}
\arguments{
\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{jobId}{a character string or a vector containing job ids.}

\item{action}{action to be performed on the jobs}

\item{params}{list containing additional query or body params}

\item{memberId}{a character or a vector contaning user or group ids to 
work on. Mandatory when using jobAclClient}
}
\description{
This function implements the OpenCGA calls for managing Jobs
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
