% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestResponse-methods.R
\name{mergeResponses}
\alias{mergeResponses}
\title{Merge a list of RestResponses}
\usage{
mergeResponses(restResponseList)
}
\arguments{
\item{RestResponseList}{a list of RestResponse objects obtained after executing 
a query to OpenCGA.}
}
\value{
a merged RestResponse object
}
\description{
Fetch the list of results of the response specified in _response_.
}
