% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestResponse-methods.R
\name{mergeResults}
\alias{mergeResults}
\title{Merge a list of RestResponse Results}
\usage{
mergeResults(resultsList)
}
\arguments{
\item{resultsList}{a list of results from a RestResponse object}
}
\value{
a character string containing the JSON with the merged results
}
\description{
Merge a list of results from multiple calls to the db
}
