% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta-methods.R
\name{metaClient,OpencgaR-method}
\alias{metaClient,OpencgaR-method}
\title{MetaClient methods}
\usage{
\S4method{metaClient}{OpencgaR}(OpencgaR, endpointName, params = NULL, ...)
}
\arguments{
\item{category}{List of categories to get API from.}
}
\description{
This function implements the OpenCGA calls for managing Meta.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   about \tab /{apiVersion}/meta/about \tab  \cr
   api \tab /{apiVersion}/meta/api \tab category \cr
   fail \tab /{apiVersion}/meta/fail \tab  \cr
   ping \tab /{apiVersion}/meta/ping \tab  \cr
   status \tab /{apiVersion}/meta/status \tab  \cr
}
}
\section{Endpoint /{apiVersion}/meta/about}{

Returns info about current OpenCGA code.
}

\section{Endpoint /{apiVersion}/meta/api}{

API.
}

\section{Endpoint /{apiVersion}/meta/fail}{

Ping Opencga webservices.
}

\section{Endpoint /{apiVersion}/meta/ping}{

Ping Opencga webservices.
}

\section{Endpoint /{apiVersion}/meta/status}{

Database status.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
