% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaClasses.R
\docType{methods}
\name{OpencgaCohorts,Opencga-method}
\alias{OpencgaCohorts,Opencga-method}
\alias{OpencgaCohorts}
\title{A method to query Opencga cohorts}
\usage{
\S4method{OpencgaCohorts}{Opencga}(object, id, action, params, ...)
}
\description{
This method allow the user to create, update and explore
cohorts data and metadta
}
