% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaR-methods.R
\name{OpencgaConfig}
\alias{OpencgaConfig}
\alias{OpencgaR}
\title{OpencgaR}
\usage{
OpencgaConfig(host, version)
}
\arguments{
\item{host}{a character specifying the host url, e.g.
"http://bioinfo.hpc.cam.ac.uk/hgva"}

\item{version}{a character specifying the API version, e.g. "v1"}
}
\value{
An object of class OpencgaR
}
\description{
This is a constructor function for the OpencgaR object
}
\details{
OpencgaR constructor function


This class defines the OpencgaR object. It holds the default configuration 
required by OpencgaR methods to connect to OpenCGA web services. By defult, 
it is configured to query HGVA (http://hgva.opencb.org/).
}
\examples{
   cga <- OpencgaR()
   print(cga)
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} 
and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
