% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaClasses.R
\docType{methods}
\name{OpencgaFiles,Opencga-method}
\alias{OpencgaFiles,Opencga-method}
\alias{OpencgaFiles}
\title{A method to query Opencga Files}
\usage{
\S4method{OpencgaFiles}{Opencga}(object, id = NULL, action, params = NULL,
  ...)
}
\description{
This method allow the user to create, update and explore
files data and metadta
}
