% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaR-methods.R
\name{opencgaHelp}
\alias{opencgaHelp}
\alias{OpencgaHelp}
\title{Get help from OpenCGA Web Services}
\usage{
opencgaHelp(opencga, client = NULL, action = NULL)
}
\arguments{
\item{opencga}{an object of type OpencgaR generated using initOpencga or 
opencgaLogin}
}
\description{
A function extract all the available information from Opencga web services 
at the defined host
}
\examples{
con <- initOpencgaR(host = "http://localhost:8080/opencga/", version = "v1")
con <- opencgaLogin(opencga = con, userid = "user", passwd = "user_pass")
opencgaHelp(con, client="userClient", action="info")

}
