% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaClasses.R
\docType{methods}
\name{OpencgaIndividuals,Opencga-method}
\alias{OpencgaIndividuals,Opencga-method}
\alias{OpencgaIndividuals}
\title{A method to query Opencga Individuals}
\usage{
\S4method{OpencgaIndividuals}{Opencga}(object, id = NULL, action, params, ...)
}
\description{
This method allow the user to create, update and explore
Individuals data and metadta
}
