% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaClasses.R
\docType{methods}
\name{OpencgaJobs,Opencga-method}
\alias{OpencgaJobs,Opencga-method}
\alias{OpencgaJobs}
\title{A method to query Opencga Jobs}
\usage{
\S4method{OpencgaJobs}{Opencga}(object, id, action, params, ...)
}
\description{
This method allow the user to create, update and explore
jobs data and metadta
}
