% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaFunctions.R
\name{OpencgaLogin}
\alias{OpencgaLogin}
\title{A function to login Opencga web services}
\usage{
OpencgaLogin(baseurl, userid = NULL, passwd = NULL, interactive = FALSE)
}
\arguments{
\item{baseurl}{a character specifying the host url}

\item{userid}{a charatcer the username}

\item{passwd}{a charcter the user password}
}
\value{
an Opencga class object
}
\description{
A function to login Opencga web services
}

