% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaR-methods.R
\name{opencgaLogout}
\alias{opencgaLogout}
\alias{OpencgaLogout}
\title{Logout from OpenCGA Web Services}
\usage{
opencgaLogout(opencga)
}
\arguments{
\item{ocga}{an object of type OpencgaR generated using opencgaLogin}
}
\description{
A function to logout from Opencga web services
}
\examples{
opencgaLogout(con)

}
