% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaClasses.R
\name{OpencgaParam}
\alias{OpencgaParam}
\title{A class to store Opencga parameters for filtering queries}
\usage{
OpencgaParam(region = character(), chromosome = character(),
  type = character(), gene = character(), maf = character(),
  mgf = character(), genotype = character(), polyphen = character(),
  sift = character(), conservation = character(), reference = character(),
  alternate = character(), so = character(), biotype = character(),
  skip = numeric(), limit = numeric(), files = character())
}
\arguments{
\item{chromosome}{the name of the chromosome to be returned}
}

