% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaClasses.R
\docType{methods}
\name{OpencgaProjects,Opencga-method}
\alias{OpencgaProjects,Opencga-method}
\alias{OpencgaProjects}
\title{A method to query Opencga Projects}
\usage{
\S4method{OpencgaProjects}{Opencga}(object, id = NULL, action,
  params = NULL, ...)
}
\description{
This method allow the user to create, update and explore
projects data and metadta
}
\details{
A method to query, and manipulate Projects data
}
