% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaR-methods.R
\name{opencgaReadConfig}
\alias{opencgaReadConfig}
\title{Read OpenCGA configuration}
\usage{
opencgaReadConfig(conf)
}
\arguments{
\item{conf}{a list or the path to a file (in "JSON" or "YAML" format) 
containing the host and version configurations to set OpenCGA connection}
}
\value{
a OpencgaR object
}
\description{
Read OpenCGA configuration
}
\examples{
# Configuration in list format
conf <- list(version="v1",
             rest=list(host="http://localhost:8080/opencga/"))
con <- OpencgaReadConfig(conf)

# Configuration in file format ("YAML" or "JSON")
conf <- "/path/to/conf/client-configuration.yml"
con <- OpencgaReadConfig(conf)

}
