% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaClasses.R
\docType{methods}
\name{OpencgaSamples,Opencga-method}
\alias{OpencgaSamples,Opencga-method}
\alias{OpencgaSamples}
\title{A method to query Opencga Samples}
\usage{
\S4method{OpencgaSamples}{Opencga}(object, id = NULL, action, params = NULL,
  ...)
}
\description{
This method allow the user to create, update and explore
Samples data and metadta
}
