% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaClasses.R
\docType{methods}
\name{OpencgaStudy,Opencga-method}
\alias{OpencgaStudy,Opencga-method}
\alias{OpencgaStudy}
\title{A method to query Opencga studies}
\usage{
\S4method{OpencgaStudy}{Opencga}(object, id = NULL, action, params = NULL,
  ...)
}
\description{
This method allow the user to create, update and explore
study data and metadta
}
\details{
A method to query Studies
}
