% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaClasses.R
\docType{methods}
\name{OpencgaUser,Opencga-method}
\alias{OpencgaUser,Opencga-method}
\alias{OpencgaUser}
\title{A method to query user data}
\usage{
\S4method{OpencgaUser}{Opencga}(object, category, id = NULL, action,
  params = NULL, ...)
}
\description{
This method allow the user to create, update, and explore user
 data and metadata
}
\details{
A method to query user data
}
