% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaClasses.R
\docType{methods}
\name{OpencgaVariables,Opencga-method}
\alias{OpencgaVariables,Opencga-method}
\alias{OpencgaVariables}
\title{A method to query Opencga Variables}
\usage{
\S4method{OpencgaVariables}{Opencga}(object, id, action, params, ...)
}
\description{
This method allow the user to create, update and explore
vaiables data and metadta
}
