% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TokenRenewal-methods.R
\docType{methods}
\name{renewToken,OpencgaR-method}
\alias{renewToken,OpencgaR-method}
\title{Token renewal method}
\usage{
\S4method{renewToken}{OpencgaR}(OpencgaR, host, version, params = NULL,
  ...)
}
\description{
This function renews a token
}
\section{Slots}{

\describe{
\item{\code{OpencgaR}}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{\code{host}}{the host}

\item{\code{version}}{the version}

\item{\code{params}}{list containing additional query or body params}
}}

\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
