% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RestResponse-class}
\alias{RestResponse-class}
\alias{restResponse}
\title{RestResponse Class}
\description{
This is an internal S4 class which defines the RestResponse object
}
\details{
REST web services return the response wrapped in a RestResponse 
object. This consists of some metadata and a list of _OpenCGAResult_ objects 
called _responses_ containing the data results and metadata requested. 
The *first response of the list* will always contain the response of the 
OpenCGA federation being directly queried. Any additional response in the 
list will belong to other federated servers that could be connected. Each 
federated response will contain a list of results ( _OpenCGAResult_ ) 
containing the data that has been queried.
}
\section{Slots}{

\describe{
\item{\code{apiVersion}}{a character specifying the API version. Example: "v2"}

\item{\code{time}}{integer especifying the time in milliseconds (ms) that took to 
retrieve the information. If the query requires multiple calls, time is added.}

\item{\code{events}}{list containing any warnings or errors}

\item{\code{params}}{list containg the specified parameters}

\item{\code{responses}}{list containing the results from the query. There is one list 
per OpenCGA federated servers that could be connected. The first response of 
the list will always contain the response of the OpenCGA federation being 
directly queried.}
}}

\seealso{
\url{http://docs.opencb.org/display/opencga/RESTful+Web+Services#RESTfulWebServices-RESTResponse}
}
