% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestResponse-methods.R
\name{results}
\alias{results}
\title{Get results from RestResponse}
\usage{
results(restResponse, response = 1)
}
\arguments{
\item{response}{Position of the response from the array of responses. Default: 1}

\item{RestResponse}{an object of class RestResponse obtained after executing 
a query to OpenCGA.}
}
\value{
the list of results of the specified response. If no _response_ is 
given, the results from the first list are returned.
}
\description{
Fetch the list of results of the response specified in _response_.
}
