% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-methods.R
\docType{methods}
\name{sampleClient,OpencgaR-method}
\alias{sampleClient,OpencgaR-method}
\title{SampleClient methods}
\usage{
\S4method{sampleClient}{OpencgaR}(OpencgaR, sample, action, params = NULL)
}
\arguments{
\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{sample}{a character string or a vector containing sample IDs or names}

\item{action}{action to be performed on the sample(s)}

\item{params}{list containing additional query or body params}

\item{annotationsetName}{a character string with the annotationset name. Only 
necessary when updating and deleting using the sampleAnnotationClient}

\item{memberId}{a character or a vector contaning user or group ids to 
work on. Mandatory when using sampleAclClient}
}
\description{
This function implements the OpenCGA calls for managing Samples
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
