% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study-methods.R
\name{studyClient,OpencgaR-method}
\alias{studyClient,OpencgaR-method}
\title{StudyClient methods}
\usage{
\S4method{studyClient}{OpencgaR}(
  OpencgaR,
  group,
  id,
  members,
  studies,
  study,
  templateId,
  variableSet,
  endpointName,
  params = NULL,
  ...
)
}
\arguments{
\item{group}{Group name.}

\item{id}{Id of the variableSet to be retrieved. If no id is passed, it will show all the variableSets of the study.}

\item{members}{Comma separated list of user or group ids.}

\item{studies}{Comma separated list of Studies [\link{organization@}project:]study where study and project can be either the ID or UUID up to a maximum of 100.}

\item{study}{Study [\link{organization@}project:]study where study and project can be either the ID or UUID.}

\item{templateId}{Template id.}

\item{variableSet}{VariableSet id of the VariableSet to be updated.}

\item{project}{Project \link{organization@}project where project can be either the ID or the alias.}

\item{name}{Study name.}

\item{alias}{Study alias.}

\item{fqn}{Study full qualified name.}

\item{internalStatus}{Filter by internal status.}

\item{attributes}{Attributes.}

\item{release}{Release value.}

\item{member}{User or group id.}

\item{limit}{Number of results to be returned.}

\item{skip}{Number of results to skip.}

\item{count}{Get the total number of results matching the query. Deactivated by default.}

\item{operationId}{Audit operation UUID.}

\item{resource}{Resource involved. Allowed values: \link{'AUDIT NOTE ORGANIZATION USER PROJECT STUDY FILE SAMPLE JOB INDIVIDUAL COHORT DISEASE_PANEL FAMILY CLINICAL_ANALYSIS INTERPRETATION VARIANT ALIGNMENT CLINICAL EXPRESSION RGA FUNCTIONAL WORKFLOW RESOURCE'}}

\item{resourceId}{Resource ID.}

\item{resourceUuid}{resource UUID.}

\item{status}{Filter by status. Allowed values: \link{'SUCCESS ERROR'}}

\item{date}{Date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{silent}{Boolean to retrieve all possible entries that are queried for, false to raise an exception whenever one of the entries looked for cannot be shown for whichever reason.}

\item{creationDate}{Creation date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{modificationDate}{Modification date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{type}{Note type.}

\item{uuid}{Unique 32-character identifier assigned automatically by OpenCGA.}

\item{userId}{User that wrote that Note.}

\item{tags}{Note tags.}

\item{visibility}{Visibility of the Note.}

\item{version}{Autoincremental version assigned to the registered entry. By default, updates does not create new versions. To enable versioning, users must set the \code{incVersion} flag from the /update web service when updating the document.}

\item{tagsAction}{Action to be performed if the array of tags is being updated. Allowed values: \link{'ADD SET REMOVE'}}

\item{entity}{Entity where the permission rules should be applied to. Allowed values: \link{'SAMPLES FILES COHORTS INDIVIDUALS FAMILIES JOBS CLINICAL_ANALYSES DISEASE_PANELS'}}

\item{jobId}{Job ID. It must be a unique string within the study. An ID will be autogenerated automatically if not provided.}

\item{jobDependsOn}{Comma separated list of existing job IDs the job will depend on.}

\item{jobDescription}{Job description.}

\item{jobTags}{Job tags.}

\item{jobScheduledStartTime}{Time when the job is scheduled to start.}

\item{jobPriority}{Priority of the job.}

\item{jobDryRun}{Flag indicating that the job will be executed in dry-run mode. In this mode, OpenCGA will validate that all parameters and prerequisites are correctly set for successful execution, but the job will not actually run.}

\item{file}{File to upload.}

\item{include}{Fields included in the response, whole JSON path must be provided.}

\item{exclude}{Fields excluded in the response, whole JSON path must be provided.}

\item{includeResult}{Flag indicating to include the created or updated document result in the response.}

\item{action}{Action to be performed: ADD or REMOVE a variable. Allowed values: \link{'ADD REMOVE'}}

\item{data}{JSON containing the variable to be added or removed. For removing, only the variable id will be needed.}
}
\description{
This function implements the OpenCGA calls for managing Studies.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   updateAcl \tab /{apiVersion}/studies/acl/{members}/update \tab members\link{*}, action\link{*}, body\link{*} \cr
   create \tab /{apiVersion}/studies/create \tab include, exclude, project, includeResult, body\link{*} \cr
   search \tab /{apiVersion}/studies/search \tab include, exclude, limit, skip, count, project\link{*}, name, id, alias, fqn, creationDate, modificationDate, internalStatus, status, attributes, release \cr
   acl \tab /{apiVersion}/studies/{studies}/acl \tab studies\link{*}, member, silent \cr
   info \tab /{apiVersion}/studies/{studies}/info \tab include, exclude, studies\link{*} \cr
   searchAudit \tab /{apiVersion}/studies/{study}/audit/search \tab include, exclude, limit, skip, count, study\link{*}, operationId, userId, action, resource, resourceId, resourceUuid, status, date \cr
   groups \tab /{apiVersion}/studies/{study}/groups \tab study\link{*}, id, silent \cr
   updateGroups \tab /{apiVersion}/studies/{study}/groups/update \tab study\link{*}, action, body\link{*} \cr
   updateGroupsUsers \tab /{apiVersion}/studies/{study}/groups/{group}/users/update \tab study\link{*}, group\link{*}, action, body\link{*} \cr
   createNotes \tab /{apiVersion}/studies/{study}/notes/create \tab include, exclude, study\link{*}, includeResult, body\link{*} \cr
   searchNotes \tab /{apiVersion}/studies/{study}/notes/search \tab include, exclude, study\link{*}, creationDate, modificationDate, id, type, uuid, userId, tags, visibility, version \cr
   deleteNotes \tab /{apiVersion}/studies/{study}/notes/{id}/delete \tab study\link{*}, id\link{*}, includeResult \cr
   updateNotes \tab /{apiVersion}/studies/{study}/notes/{id}/update \tab include, exclude, study\link{*}, id\link{*}, tagsAction, includeResult, body\link{*} \cr
   permissionRules \tab /{apiVersion}/studies/{study}/permissionRules \tab study\link{*}, entity\link{*} \cr
   updatePermissionRules \tab /{apiVersion}/studies/{study}/permissionRules/update \tab study\link{*}, entity\link{*}, action, body\link{*} \cr
   runTemplates \tab /{apiVersion}/studies/{study}/templates/run \tab study\link{*}, jobId, jobDependsOn, jobDescription, jobTags, jobScheduledStartTime, jobPriority, jobDryRun, body\link{*} \cr
   uploadTemplates \tab /{apiVersion}/studies/{study}/templates/upload \tab file, study\link{*} \cr
   deleteTemplates \tab /{apiVersion}/studies/{study}/templates/{templateId}/delete \tab study\link{*}, templateId\link{*} \cr
   update \tab /{apiVersion}/studies/{study}/update \tab include, exclude, study\link{*}, includeResult, body\link{*} \cr
   variableSets \tab /{apiVersion}/studies/{study}/variableSets \tab study\link{*}, id \cr
   updateVariableSets \tab /{apiVersion}/studies/{study}/variableSets/update \tab study\link{*}, action, body\link{*} \cr
   updateVariableSetsVariables \tab /{apiVersion}/studies/{study}/variableSets/{variableSet}/variables/update \tab study\link{*}, variableSet\link{*}, action, body\link{*} \cr
}
}
\section{Endpoint /{apiVersion}/studies/acl/{members}/update}{

Update the set of permissions granted for the member.
}

\section{Endpoint /{apiVersion}/studies/create}{

Create a new study.
}

\section{Endpoint /{apiVersion}/studies/search}{

Search studies.
}

\section{Endpoint /{apiVersion}/studies/{studies}/acl}{

Return the acl of the study. If member is provided, it will only return the acl for the member.
}

\section{Endpoint /{apiVersion}/studies/{studies}/info}{

Fetch study information.
}

\section{Endpoint /{apiVersion}/studies/{study}/audit/search}{

Search audit collection.
}

\section{Endpoint /{apiVersion}/studies/{study}/groups}{

Return the groups present in the study. For owners and administrators only.
}

\section{Endpoint /{apiVersion}/studies/{study}/groups/update}{

Add or remove a group.
}

\section{Endpoint /{apiVersion}/studies/{study}/groups/{group}/users/update}{

Add, set or remove users from an existing group.
}

\section{Endpoint /{apiVersion}/studies/{study}/notes/create}{

Create a new note.
}

\section{Endpoint /{apiVersion}/studies/{study}/notes/search}{

Search for notes of scope STUDY.
}

\section{Endpoint /{apiVersion}/studies/{study}/notes/{id}/delete}{

Delete note.
}

\section{Endpoint /{apiVersion}/studies/{study}/notes/{id}/update}{

Update a note.
}

\section{Endpoint /{apiVersion}/studies/{study}/permissionRules}{

Fetch permission rules.
}

\section{Endpoint /{apiVersion}/studies/{study}/permissionRules/update}{

Add or remove a permission rule.
}

\section{Endpoint /{apiVersion}/studies/{study}/templates/run}{

Execute template.
}

\section{Endpoint /{apiVersion}/studies/{study}/templates/upload}{

Resource to upload a zipped template.
}

\section{Endpoint /{apiVersion}/studies/{study}/templates/{templateId}/delete}{

Delete template.
}

\section{Endpoint /{apiVersion}/studies/{study}/update}{

Update some study attributes.
}

\section{Endpoint /{apiVersion}/studies/{study}/variableSets}{

Fetch variableSets from a study.
}

\section{Endpoint /{apiVersion}/studies/{study}/variableSets/update}{

Add or remove a variableSet.
}

\section{Endpoint /{apiVersion}/studies/{study}/variableSets/{variableSet}/variables/update}{

Add or remove variables to a VariableSet.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
