% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow-methods.R
\name{workflowClient,OpencgaR-method}
\alias{workflowClient,OpencgaR-method}
\title{WorkflowClient methods}
\usage{
\S4method{workflowClient}{OpencgaR}(
  OpencgaR,
  members,
  workflowId,
  workflows,
  endpointName,
  params = NULL,
  ...
)
}
\arguments{
\item{members}{Comma separated list of user or group ids.}

\item{workflowId}{Comma separated list workflow IDs or UUIDs up to a maximum of 100.}

\item{workflows}{Comma separated list sample IDs or UUIDs up to a maximum of 100.}

\item{action}{Action to be performed \link{ADD, SET, REMOVE or RESET}. Allowed values: \link{'SET ADD REMOVE RESET'}}

\item{field}{Comma separated list of fields for which to obtain the distinct values.}

\item{jobId}{Job ID. It must be a unique string within the study. An ID will be autogenerated automatically if not provided.}

\item{jobDescription}{Job description.}

\item{jobDependsOn}{Comma separated list of existing job IDs the job will depend on.}

\item{jobTags}{Job tags.}

\item{jobScheduledStartTime}{Time when the job is scheduled to start.}

\item{jobPriority}{Priority of the job.}

\item{jobDryRun}{Flag indicating that the job will be executed in dry-run mode. In this mode, OpenCGA will validate that all parameters and prerequisites are correctly set for successful execution, but the job will not actually run.}

\item{limit}{Number of results to be returned.}

\item{skip}{Number of results to skip.}

\item{count}{Get the total number of results matching the query. Deactivated by default.}

\item{id}{Comma separated list of workflow IDs up to a maximum of 100. Also admits basic regular expressions using the operator '~', i.e. '~{perl-regex}' e.g. '~value' for case sensitive, '~/value/i' for case insensitive search.}

\item{name}{Comma separated list of workflow names up to a maximum of 100. Also admits basic regular expressions using the operator '~', i.e. '~{perl-regex}' e.g. '~value' for case sensitive, '~/value/i' for case insensitive search.}

\item{uuid}{Comma separated list of workflow UUIDs up to a maximum of 100.}

\item{tags}{Comma separated list of tags.}

\item{draft}{Boolean field indicating whether the workflow is a draft or not.}

\item{internal.registrationUserId}{UserId that created the workflow.}

\item{manager.id}{Id of the workflow system (Allowed values: NEXTFLOW).}

\item{type}{Workflow type. Allowed types: \link{CLINICAL_INTERPRETATION, SECONDARY_ANALYSIS, RESEARCH or OTHER}.}

\item{creationDate}{Creation date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{modificationDate}{Modification date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{acl}{Filter entries for which a user has the provided permissions. Format: acl={user}:{permissions}. Example: acl=john:WRITE,WRITE_ANNOTATIONS will return all entries for which user john has both WRITE and WRITE_ANNOTATIONS permissions. Only study owners or administrators can query by this field. .}

\item{release}{Release when it was created.}

\item{snapshot}{Snapshot value (Latest version of the entry in the specified release).}

\item{includeResult}{Flag indicating to include the created or updated document result in the response.}

\item{data}{body.}

\item{member}{User or group id.}

\item{silent}{Boolean to retrieve all possible entries that are queried for, false to raise an exception whenever one of the entries looked for cannot be shown for whichever reason.}

\item{include}{Fields included in the response, whole JSON path must be provided.}

\item{exclude}{Fields excluded in the response, whole JSON path must be provided.}

\item{study}{Study [\link{organization@}project:]study where study and project can be either the ID or UUID.}

\item{version}{Comma separated list of workflow versions. 'all' to get all the workflow versions. Not supported if multiple workflow ids are provided.}

\item{deleted}{Boolean to retrieve deleted entries.}
}
\description{
This function implements the OpenCGA calls for managing Workflows.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   updateAcl \tab /{apiVersion}/workflows/acl/{members}/update \tab study, members\link{*}, action\link{*}, body\link{*} \cr
   create \tab /{apiVersion}/workflows/create \tab include, exclude, study, includeResult, body\link{*} \cr
   distinct \tab /{apiVersion}/workflows/distinct \tab study, id, name, uuid, tags, draft, internal.registrationUserId, manager.id, type, creationDate, modificationDate, acl, release, snapshot, deleted, field\link{*} \cr
   importWorkflow \tab /{apiVersion}/workflows/import \tab study, body\link{*} \cr
   run \tab /{apiVersion}/workflows/run \tab study, jobId, jobDescription, jobDependsOn, jobTags, jobScheduledStartTime, jobPriority, jobDryRun, body\link{*} \cr
   search \tab /{apiVersion}/workflows/search \tab include, exclude, limit, skip, count, study, id, name, uuid, tags, draft, internal.registrationUserId, manager.id, type, creationDate, modificationDate, acl, release, snapshot, deleted \cr
   update \tab /{apiVersion}/workflows/{workflowId}/update \tab include, exclude, workflowId\link{*}, study, includeResult, body \cr
   acl \tab /{apiVersion}/workflows/{workflows}/acl \tab workflows\link{*}, study, member, silent \cr
   delete \tab /{apiVersion}/workflows/{workflows}/delete \tab study, workflows\link{*} \cr
   info \tab /{apiVersion}/workflows/{workflows}/info \tab include, exclude, workflows\link{*}, study, version, deleted \cr
}
}
\section{Endpoint /{apiVersion}/workflows/acl/{members}/update}{

Update the set of workflow permissions granted for the member.
}

\section{Endpoint /{apiVersion}/workflows/create}{

Create a workflow.
}

\section{Endpoint /{apiVersion}/workflows/distinct}{

Workflow distinct method.
}

\section{Endpoint /{apiVersion}/workflows/import}{

Import a workflow.
}

\section{Endpoint /{apiVersion}/workflows/run}{

Execute a workflow analysis.
}

\section{Endpoint /{apiVersion}/workflows/search}{

Workflow search method.
}

\section{Endpoint /{apiVersion}/workflows/{workflowId}/update}{

Update some workflow attributes.
}

\section{Endpoint /{apiVersion}/workflows/{workflows}/acl}{

Returns the acl of the workflows. If member is provided, it will only return the acl for the member.
}

\section{Endpoint /{apiVersion}/workflows/{workflows}/delete}{

Delete workflows.
}

\section{Endpoint /{apiVersion}/workflows/{workflows}/info}{

Get workflow information.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
