% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Admin-methods.R
\name{adminClient,OpencgaR-method}
\alias{adminClient,OpencgaR-method}
\title{AdminClient methods}
\usage{
\S4method{adminClient}{OpencgaR}(OpencgaR, endpointName, params = NULL, ...)
}
\arguments{
\item{count}{Count the number of elements matching the group.}

\item{limit}{Maximum number of documents (groups) to be returned.}

\item{fields}{Comma separated list of fields by which to group by.}

\item{entity}{Entity to be grouped by.}

\item{action}{Action performed.}

\item{before}{Object before update.}

\item{after}{Object after update.}

\item{date}{Date <,<=,>,>=(Format: yyyyMMddHHmmss) and yyyyMMddHHmmss-yyyyMMddHHmmss.}

\item{data}{JSON containing the parameters.}
}
\description{
This function implements the OpenCGA calls for managing Admin.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   groupByAudit \tab /{apiVersion}/admin/audit/groupBy \tab count, limit, fields\link{*}, entity\link{*}, action, before, after, date \cr
   indexStatsCatalog \tab /{apiVersion}/admin/catalog/indexStats \tab  \cr
   installCatalog \tab /{apiVersion}/admin/catalog/install \tab body\link{*} \cr
   jwtCatalog \tab /{apiVersion}/admin/catalog/jwt \tab body\link{*} \cr
   createUsers \tab /{apiVersion}/admin/users/create \tab body\link{*} \cr
   importUsers \tab /{apiVersion}/admin/users/import \tab body\link{*} \cr
   syncUsers \tab /{apiVersion}/admin/users/sync \tab body\link{*} \cr
}
}
\section{Endpoint /{apiVersion}/admin/audit/groupBy}{

Group by operation.
}

\section{Endpoint /{apiVersion}/admin/catalog/indexStats}{

Sync Catalog into the Solr.
}

\section{Endpoint /{apiVersion}/admin/catalog/install}{

Install OpenCGA database.
}

\section{Endpoint /{apiVersion}/admin/catalog/jwt}{

Change JWT secret key.
}

\section{Endpoint /{apiVersion}/admin/users/create}{

Create a new user.
}

\section{Endpoint /{apiVersion}/admin/users/import}{

Import users or a group of users from LDAP or AAD.
}

\section{Endpoint /{apiVersion}/admin/users/sync}{

Synchronise groups of users with LDAP groups.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
