% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Admin-methods.R
\name{adminClient,OpencgaR-method}
\alias{adminClient,OpencgaR-method}
\title{AdminClient methods}
\usage{
\S4method{adminClient}{OpencgaR}(OpencgaR, endpointName, params = NULL, ...)
}
\arguments{
\item{fields}{Comma separated list of fields by which to group by.}

\item{entity}{Entity to be grouped by.}

\item{action}{Action performed.}

\item{before}{Object before update.}

\item{after}{Object after update.}

\item{date}{Date <,<=,>,>=(Format: yyyyMMddHHmmss) and yyyyMMddHHmmss-yyyyMMddHHmmss.}

\item{collection}{Collection to be indexed (file, sample, individual, family, cohort and/or job). If not provided, all of them will be indexed.}

\item{include}{Fields included in the response, whole JSON path must be provided.}

\item{exclude}{Fields excluded in the response, whole JSON path must be provided.}

\item{limit}{Number of results to be returned.}

\item{skip}{Number of results to skip.}

\item{count}{Get the total number of results matching the query. Deactivated by default.}

\item{user}{User ID.}

\item{account}{Account type \link{GUEST, FULL, ADMINISTRATOR}.}

\item{authenticationId}{Authentication origin ID.}

\item{data}{JSON containing the parameters.}
}
\description{
This function implements the OpenCGA calls for managing Admin.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   groupByAudit \tab /{apiVersion}/admin/audit/groupBy \tab count, limit, fields\link{*}, entity\link{*}, action, before, after, date \cr
   indexStatsCatalog \tab /{apiVersion}/admin/catalog/indexStats \tab collection \cr
   installCatalog \tab /{apiVersion}/admin/catalog/install \tab body\link{*} \cr
   jwtCatalog \tab /{apiVersion}/admin/catalog/jwt \tab body\link{*} \cr
   createUsers \tab /{apiVersion}/admin/users/create \tab body\link{*} \cr
   importUsers \tab /{apiVersion}/admin/users/import \tab body\link{*} \cr
   searchUsers \tab /{apiVersion}/admin/users/search \tab include, exclude, limit, skip, count, user, account, authenticationId \cr
   syncUsers \tab /{apiVersion}/admin/users/sync \tab body\link{*} \cr
}
}
\section{Endpoint /{apiVersion}/admin/audit/groupBy}{

Group by operation.
}

\section{Endpoint /{apiVersion}/admin/catalog/indexStats}{

Sync Catalog into the Solr.
}

\section{Endpoint /{apiVersion}/admin/catalog/install}{

Install OpenCGA database.
}

\section{Endpoint /{apiVersion}/admin/catalog/jwt}{

Change JWT secret key.
}

\section{Endpoint /{apiVersion}/admin/users/create}{

Create a new user.
}

\section{Endpoint /{apiVersion}/admin/users/import}{

Import users or a group of users from LDAP or AAD.
}

\section{Endpoint /{apiVersion}/admin/users/search}{

User search method.
}

\section{Endpoint /{apiVersion}/admin/users/sync}{

Synchronise a group of users from an authentication origin with a group in a study from catalog.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
