% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clinical-methods.R
\docType{methods}
\name{clinicalClient,OpencgaR-method}
\alias{clinicalClient,OpencgaR-method}
\title{ClinicalClient methods}
\usage{
\S4method{clinicalClient}{OpencgaR}(OpencgaR, clinicalAnalysis, action,
  params = NULL, ...)
}
\arguments{
\item{OpencgaR}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{clinicalAnalysis}{a character string or vector containing clinical analysis IDs}

\item{action}{action to be performed on the sample(s)}

\item{params}{list containing additional query or body params}
}
\description{
This function implements the OpenCGA calls for managing Clinical Analyses
}
\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
