% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clinical-methods.R
\name{clinicalClient,OpencgaR-method}
\alias{clinicalClient,OpencgaR-method}
\title{ClinicalClient methods}
\usage{
\S4method{clinicalClient}{OpencgaR}(
  OpencgaR,
  clinicalAnalyses,
  clinicalAnalysis,
  members,
  endpointName,
  params = NULL,
  ...
)
}
\arguments{
\item{clinicalAnalyses}{Comma separated list of clinical analysis IDs.}

\item{clinicalAnalysis}{Clinical analysis ID.}

\item{members}{Comma separated list of user or group IDs.}

\item{sort}{Sort the results.}

\item{software}{Software.}

\item{analyst}{Analyst.}

\item{comment}{Comments.}

\item{version}{Version.}

\item{jobId}{Job ID. It must be a unique string within the study. An ID will be autogenerated automatically if not provided.}

\item{jobDescription}{Job description.}

\item{jobDependsOn}{Comma separated list of existing job IDs the job will depend on.}

\item{jobTags}{Job tags.}

\item{priority}{Priority.}

\item{status}{Clinical analysis status.}

\item{creationDate}{Creation date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{modificationDate}{Modification date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{dueDate}{Due date (Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805...).}

\item{description}{Description.}

\item{proband}{Proband.}

\item{analystAssignee}{Clinical analyst assignee.}

\item{disorder}{Disorder ID or name.}

\item{flags}{Flags.}

\item{release}{Release value.}

\item{attributes}{Text attributes (Format: sex=male,age>20 ...).}

\item{limit}{Number of results to be returned.}

\item{skip}{Number of results to skip.}

\item{count}{Get the total number of results matching the query. Deactivated by default.}

\item{approximateCount}{Get an approximate count, instead of an exact total count. Reduces execution time.}

\item{approximateCountSamplingSize}{Sampling size to get the approximate count. Larger values increase accuracy but also increase execution time.}

\item{savedFilter}{Use a saved filter at User level.}

\item{id}{List of IDs, these can be rs IDs (dbSNP) or variants in the format chrom:start:ref:alt, e.g. rs116600158,19:7177679:C:T.}

\item{region}{List of regions, these can be just a single chromosome name or regions in the format chr:start-end, e.g.: 2,3:100000-200000.}

\item{type}{List of types, accepted values are SNV, MNV, INDEL, SV, CNV, INSERTION, DELETION, e.g. SNV,INDEL.}

\item{file}{Filter variants from the files specified. This will set includeFile parameter when not provided.}

\item{filter}{Specify the FILTER for any of the files. If 'file' filter is provided, will match the file and the filter. e.g.: PASS,LowGQX.}

\item{qual}{Specify the QUAL for any of the files. If 'file' filter is provided, will match the file and the qual. e.g.: >123.4.}

\item{fileData}{Filter by file data (i.e. INFO column from VCF file). [{file}:]{key}{op}{value}\link{,;}* . If no file is specified, will use all files from "file" filter. e.g. AN>200 or file_1.vcf:AN>200;file_2.vcf:AN<10 . Many INFO fields can be combined. e.g. file_1.vcf:AN>200;DB=true;file_2.vcf:AN<10.}

\item{sample}{Filter variants by sample genotype. This will automatically set 'includeSample' parameter when not provided. This filter accepts multiple 3 forms: 1) List of samples: Samples that contain the main variant. Accepts AND (;) and OR (,) operators.  e.g. HG0097,HG0098 . 2) List of samples with genotypes: {sample}:{gt1},{gt2}. Accepts AND (;) and OR (,) operators.  e.g. HG0097:0/0;HG0098:0/1,1/1 . Unphased genotypes (e.g. 0/1, 1/1) will also include phased genotypes (e.g. 0|1, 1|0, 1|1), but not vice versa. When filtering by multi-allelic genotypes, any secondary allele will match, regardless of its position e.g. 1/2 will match with genotypes 1/2, 1/3, 1/4, .... Genotype aliases accepted: HOM_REF, HOM_ALT, HET, HET_REF, HET_ALT and MISS  e.g. HG0097:HOM_REF;HG0098:HET_REF,HOM_ALT . 3) Sample with segregation mode: {sample}:{segregation}. Only one sample accepted.Accepted segregation modes: \link{ monoallelic, monoallelicIncompletePenetrance, biallelic, biallelicIncompletePenetrance, XlinkedBiallelic, XlinkedMonoallelic, Ylinked, MendelianError, DeNovo, CompoundHeterozygous }. Value is case insensitive. e.g. HG0097:DeNovo Sample must have parents defined and indexed. .}

\item{sampleData}{Filter by any SampleData field from samples. [{sample}:]{key}{op}{value}\link{,;}* . If no sample is specified, will use all samples from "sample" or "genotype" filter. e.g. DP>200 or HG0097:DP>200,HG0098:DP<10 . Many FORMAT fields can be combined. e.g. HG0097:DP>200;GT=1/1,0/1,HG0098:DP<10.}

\item{sampleAnnotation}{Selects some samples using metadata information from Catalog. e.g. age>20;phenotype=hpo:123,hpo:456;name=smith.}

\item{cohort}{Select variants with calculated stats for the selected cohorts.}

\item{cohortStatsRef}{Reference Allele Frequency: [{study:}]{cohort}[<|>|<=|>=]{number}. e.g. ALL<=0.4.}

\item{cohortStatsAlt}{Alternate Allele Frequency: [{study:}]{cohort}[<|>|<=|>=]{number}. e.g. ALL<=0.4.}

\item{cohortStatsMaf}{Minor Allele Frequency: [{study:}]{cohort}[<|>|<=|>=]{number}. e.g. ALL<=0.4.}

\item{cohortStatsMgf}{Minor Genotype Frequency: [{study:}]{cohort}[<|>|<=|>=]{number}. e.g. ALL<=0.4.}

\item{cohortStatsPass}{Filter PASS frequency: [{study:}]{cohort}[<|>|<=|>=]{number}. e.g. ALL>0.8.}

\item{missingAlleles}{Number of missing alleles: [{study:}]{cohort}[<|>|<=|>=]{number}.}

\item{missingGenotypes}{Number of missing genotypes: [{study:}]{cohort}[<|>|<=|>=]{number}.}

\item{score}{Filter by variant score: [{study:}]{score}[<|>|<=|>=]{number}.}

\item{family}{Filter variants where any of the samples from the given family contains the variant (HET or HOM_ALT).}

\item{familyDisorder}{Specify the disorder to use for the family segregation.}

\item{familySegregation}{Filter by segregation mode from a given family. Accepted values: \link{ monoallelic, monoallelicIncompletePenetrance, biallelic, biallelicIncompletePenetrance, XlinkedBiallelic, XlinkedMonoallelic, Ylinked, MendelianError, DeNovo, CompoundHeterozygous }.}

\item{familyMembers}{Sub set of the members of a given family.}

\item{familyProband}{Specify the proband child to use for the family segregation.}

\item{gene}{List of genes, most gene IDs are accepted (HGNC, Ensembl gene, ...). This is an alias to 'xref' parameter.}

\item{ct}{List of SO consequence types, e.g. missense_variant,stop_lost or SO:0001583,SO:0001578.}

\item{xref}{List of any external reference, these can be genes, proteins or variants. Accepted IDs include HGNC, Ensembl genes, dbSNP, ClinVar, HPO, Cosmic, ...}

\item{biotype}{List of biotypes, e.g. protein_coding.}

\item{proteinSubstitution}{Protein substitution scores include SIFT and PolyPhen. You can query using the score {protein_score}[<|>|<=|>=]{number} or the description {protein_score}[~=|=]{description} e.g. polyphen>0.1,sift=tolerant.}

\item{conservation}{Filter by conservation score: {conservation_score}[<|>|<=|>=]{number} e.g. phastCons>0.5,phylop<0.1,gerp>0.1.}

\item{populationFrequencyAlt}{Alternate Population Frequency: {study}:{population}[<|>|<=|>=]{number}. e.g. 1kG_phase3:ALL<0.01.}

\item{populationFrequencyRef}{Reference Population Frequency: {study}:{population}[<|>|<=|>=]{number}. e.g. 1kG_phase3:ALL<0.01.}

\item{populationFrequencyMaf}{Population minor allele frequency: {study}:{population}[<|>|<=|>=]{number}. e.g. 1kG_phase3:ALL<0.01.}

\item{transcriptFlag}{List of transcript annotation flags. e.g. CCDS, basic, cds_end_NF, mRNA_end_NF, cds_start_NF, mRNA_start_NF, seleno.}

\item{geneTraitId}{List of gene trait association id. e.g. "umls:C0007222" , "OMIM:269600".}

\item{go}{List of GO (Gene Ontology) terms. e.g. "GO:0002020".}

\item{expression}{List of tissues of interest. e.g. "lung".}

\item{proteinKeyword}{List of Uniprot protein variant annotation keywords.}

\item{drug}{List of drug names.}

\item{functionalScore}{Functional score: {functional_score}[<|>|<=|>=]{number} e.g. cadd_scaled>5.2 , cadd_raw<=0.3.}

\item{clinicalSignificance}{Clinical significance: benign, likely_benign, likely_pathogenic, pathogenic.}

\item{customAnnotation}{Custom annotation: {key}[<|>|<=|>=]{number} or {key}[~=|=]{text}.}

\item{panel}{Filter by genes from the given disease panel.}

\item{trait}{List of traits, based on ClinVar, HPO, COSMIC, i.e.: IDs, histologies, descriptions,...}

\item{member}{User or group ID.}

\item{silent}{Boolean to retrieve all possible entries that are queried for, false to raise an exception whenever one of the entries looked for cannot be shown for whichever reason.}

\item{include}{Fields included in the response, whole JSON path must be provided.}

\item{exclude}{Fields excluded in the response, whole JSON path must be provided.}

\item{deleted}{Boolean to retrieve deleted entries.}

\item{study}{Study [\link{user@}project:]study where study and project can be either the ID or UUID.}

\item{action}{Action to be performed if the array of interpretations is being updated.}

\item{data}{JSON containing clinical analysis information.}
}
\description{
This function implements the OpenCGA calls for managing Analysis - Clinical.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   updateAcl \tab /{apiVersion}/analysis/clinical/acl/{members}/update \tab study, members\link{*}, action\link{*}, body\link{*} \cr
   create \tab /{apiVersion}/analysis/clinical/create \tab study, body\link{*} \cr
   searchInterpretation \tab /{apiVersion}/analysis/clinical/interpretation/search \tab include, exclude, limit, skip, sort, study, id, description, software, analyst, comment, status, creationDate, version, panel \cr
   runInterpreterCancerTiering \tab /{apiVersion}/analysis/clinical/interpreter/cancerTiering/run \tab study, jobId, jobDescription, jobDependsOn, jobTags, body\link{*} \cr
   runInterpreterTeam \tab /{apiVersion}/analysis/clinical/interpreter/team/run \tab study, jobId, jobDescription, jobDependsOn, jobTags, body\link{*} \cr
   runInterpreterTiering \tab /{apiVersion}/analysis/clinical/interpreter/tiering/run \tab study, jobId, jobDescription, jobDependsOn, jobTags, body\link{*} \cr
   runInterpreterZetta \tab /{apiVersion}/analysis/clinical/interpreter/zetta/run \tab study, jobId, jobDescription, jobDependsOn, jobTags, body\link{*} \cr
   search \tab /{apiVersion}/analysis/clinical/search \tab include, exclude, limit, skip, count, study, id, type, priority, status, creationDate, modificationDate, dueDate, description, family, proband, sample, analystAssignee, disorder, flags, deleted, release, attributes \cr
   actionableVariant \tab /{apiVersion}/analysis/clinical/variant/actionable \tab study, sample \cr
   queryVariant \tab /{apiVersion}/analysis/clinical/variant/query \tab include, exclude, limit, skip, count, approximateCount, approximateCountSamplingSize, savedFilter, id, region, type, study, file, filter, qual, fileData, sample, sampleData, sampleAnnotation, cohort, cohortStatsRef, cohortStatsAlt, cohortStatsMaf, cohortStatsMgf, cohortStatsPass, missingAlleles, missingGenotypes, score, family, familyDisorder, familySegregation, familyMembers, familyProband, gene, ct, xref, biotype, proteinSubstitution, conservation, populationFrequencyAlt, populationFrequencyRef, populationFrequencyMaf, transcriptFlag, geneTraitId, go, expression, proteinKeyword, drug, functionalScore, clinicalSignificance, customAnnotation, panel, trait \cr
   acl \tab /{apiVersion}/analysis/clinical/{clinicalAnalyses}/acl \tab clinicalAnalyses\link{*}, study, member, silent \cr
   delete \tab /{apiVersion}/analysis/clinical/{clinicalAnalyses}/delete \tab study, clinicalAnalyses \cr
   update \tab /{apiVersion}/analysis/clinical/{clinicalAnalyses}/update \tab clinicalAnalyses, study, body\link{*} \cr
   info \tab /{apiVersion}/analysis/clinical/{clinicalAnalysis}/info \tab include, exclude, clinicalAnalysis, study, deleted \cr
   updateInterpretation \tab /{apiVersion}/analysis/clinical/{clinicalAnalysis}/interpretation/update \tab study, clinicalAnalysis, body\link{*} \cr
   updateQualityControl \tab /{apiVersion}/analysis/clinical/{clinicalAnalysis}/qualityControl/update \tab study, clinicalAnalysis, body\link{*} \cr
   updateSecondaryInterpretations \tab /{apiVersion}/analysis/clinical/{clinicalAnalysis}/secondaryInterpretations/update \tab clinicalAnalysis, study, action, body\link{*} \cr
}
}
\section{Endpoint /{apiVersion}/analysis/clinical/acl/{members}/update}{

Update the set of permissions granted for the member.
}

\section{Endpoint /{apiVersion}/analysis/clinical/create}{

Create a new clinical analysis.
}

\section{Endpoint /{apiVersion}/analysis/clinical/interpretation/search}{

Clinical interpretation analysis.
}

\section{Endpoint /{apiVersion}/analysis/clinical/interpreter/cancerTiering/run}{

Run cancer tiering interpretation analysis.
}

\section{Endpoint /{apiVersion}/analysis/clinical/interpreter/team/run}{

Run TEAM interpretation analysis.
}

\section{Endpoint /{apiVersion}/analysis/clinical/interpreter/tiering/run}{

Run tiering interpretation analysis.
}

\section{Endpoint /{apiVersion}/analysis/clinical/interpreter/zetta/run}{

Run Zetta interpretation analysis.
}

\section{Endpoint /{apiVersion}/analysis/clinical/search}{

Clinical analysis search.
}

\section{Endpoint /{apiVersion}/analysis/clinical/variant/actionable}{

Fetch actionable clinical variants.
}

\section{Endpoint /{apiVersion}/analysis/clinical/variant/query}{

Fetch clinical variants.
}

\section{Endpoint /{apiVersion}/analysis/clinical/{clinicalAnalyses}/acl}{

Returns the acl of the clinical analyses. If member is provided, it will only return the acl for the member.
}

\section{Endpoint /{apiVersion}/analysis/clinical/{clinicalAnalyses}/delete}{

Delete clinical analyses.
}

\section{Endpoint /{apiVersion}/analysis/clinical/{clinicalAnalyses}/update}{

Update clinical analysis attributes.
}

\section{Endpoint /{apiVersion}/analysis/clinical/{clinicalAnalysis}/info}{

Clinical analysis info.
}

\section{Endpoint /{apiVersion}/analysis/clinical/{clinicalAnalysis}/interpretation/update}{

Update interpretation fields of primary interpretation.
}

\section{Endpoint /{apiVersion}/analysis/clinical/{clinicalAnalysis}/qualityControl/update}{

Update quality control fields of clinical analysis.
}

\section{Endpoint /{apiVersion}/analysis/clinical/{clinicalAnalysis}/secondaryInterpretations/update}{

Add or remove secondary interpretations to/from a Clinical Analysis.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
