% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort-methods.R
\name{cohortClient,OpencgaR-method}
\alias{cohortClient,OpencgaR-method}
\title{CohortClient methods}
\usage{
\S4method{cohortClient}{OpencgaR}(
  OpencgaR,
  cohorts,
  annotationSet,
  members,
  cohort,
  endpointName,
  params = NULL,
  ...
)
}
\arguments{
\item{cohorts}{Comma separated list of cohort ids.}

\item{annotationSet}{AnnotationSet ID to be updated.}

\item{members}{Comma separated list of user or group ids.}

\item{cohort}{Cohort ID.}

\item{creationYear}{Creation year.}

\item{creationMonth}{Creation month (JANUARY, FEBRUARY...).}

\item{creationDay}{Creation day.}

\item{creationDayOfWeek}{Creation day of week (MONDAY, TUESDAY...).}

\item{numSamples}{Number of samples.}

\item{default}{Calculate default stats.}

\item{field}{List of fields separated by semicolons, e.g.: studies;type. For nested fields use >>, e.g.: studies>>biotype;type;numSamples\link{0..10}:1.}

\item{variableSetId}{Variable set ID or name.}

\item{path}{Path where the TSV file is located in OpenCGA or where it should be located.}

\item{parents}{Flag indicating whether to create parent directories if they don't exist (only when TSV file was not previously associated).}

\item{annotationSetId}{Annotation set id. If not provided, variableSetId will be used.}

\item{variableSet}{Variable set ID or name.}

\item{variable}{Variable name.}

\item{limit}{Number of results to be returned.}

\item{skip}{Number of results to skip.}

\item{count}{Get the total number of results matching the query. Deactivated by default.}

\item{name}{DEPRECATED: Name of the cohort.}

\item{type}{Cohort type.}

\item{creationDate}{Creation date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{modificationDate}{Modification date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{status}{Status.}

\item{annotation}{Annotation filters. Example: age>30;gender=FEMALE. For more information, please visit http://docs.opencb.org/display/opencga/AnnotationSets+1.4.0.}

\item{acl}{Filter entries for which a user has the provided permissions. Format: acl={user}:{permissions}. Example: acl=john:WRITE,WRITE_ANNOTATIONS will return all entries for which user john has both WRITE and WRITE_ANNOTATIONS permissions. Only study owners or administrators can query by this field. .}

\item{samples}{Sample list.}

\item{release}{Release value.}

\item{member}{User or group id.}

\item{silent}{Boolean to retrieve all possible entries that are queried for, false to raise an exception whenever one of the entries looked for cannot be shown for whichever reason.}

\item{include}{Fields included in the response, whole JSON path must be provided.}

\item{exclude}{Fields excluded in the response, whole JSON path must be provided.}

\item{flattenAnnotations}{Flatten the annotations?.}

\item{deleted}{Boolean to retrieve deleted cohorts.}

\item{annotationSetsAction}{Action to be performed if the array of annotationSets is being updated. Allowed values: \link{'ADD', 'SET', 'REMOVE'}}

\item{study}{Study [\link{user@}project:]study where study and project can be either the ID or UUID.}

\item{action}{Action to be performed: ADD to add new annotations; REPLACE to replace the value of an already existing annotation; SET to set the new list of annotations removing any possible old annotations; REMOVE to remove some annotations; RESET to set some annotations to the default value configured in the corresponding variables of the VariableSet if any. Allowed values: \link{'ADD', 'SET', 'REMOVE', 'RESET', 'REPLACE'}}

\item{data}{Json containing the map of annotations when the action is ADD, SET or REPLACE, a json with only the key 'remove' containing the comma separated variables to be removed as a value when the action is REMOVE or a json with only the key 'reset' containing the comma separated variables that will be set to the default value when the action is RESET.}
}
\description{
This function implements the OpenCGA calls for managing Cohorts.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   updateAcl \tab /{apiVersion}/cohorts/acl/{members}/update \tab study, members\link{*}, action\link{*}, body\link{*} \cr
   aggregationStats \tab /{apiVersion}/cohorts/aggregationStats \tab study, type, creationYear, creationMonth, creationDay, creationDayOfWeek, numSamples, status, release, annotation, default, field \cr
   loadAnnotationSets \tab /{apiVersion}/cohorts/annotationSets/load \tab study, variableSetId\link{*}, path\link{*}, parents, annotationSetId, body \cr
   create \tab /{apiVersion}/cohorts/create \tab study, variableSet, variable, body\link{*} \cr
   search \tab /{apiVersion}/cohorts/search \tab include, exclude, limit, skip, count, flattenAnnotations, study, name, type, creationDate, modificationDate, deleted, status, annotation, acl, samples, release \cr
   acl \tab /{apiVersion}/cohorts/{cohorts}/acl \tab cohorts\link{*}, study, member, silent \cr
   delete \tab /{apiVersion}/cohorts/{cohorts}/delete \tab study, cohorts \cr
   info \tab /{apiVersion}/cohorts/{cohorts}/info \tab include, exclude, flattenAnnotations, cohorts\link{*}, study, deleted \cr
   update \tab /{apiVersion}/cohorts/{cohorts}/update \tab cohorts\link{*}, study, annotationSetsAction, body \cr
   updateAnnotations \tab /{apiVersion}/cohorts/{cohort}/annotationSets/{annotationSet}/annotations/update \tab cohort\link{*}, study, annotationSet, action, body \cr
}
}
\section{Endpoint /{apiVersion}/cohorts/acl/{members}/update}{

Update the set of permissions granted for the member.
}

\section{Endpoint /{apiVersion}/cohorts/aggregationStats}{

Fetch catalog cohort stats.
}

\section{Endpoint /{apiVersion}/cohorts/annotationSets/load}{

Load annotation sets from a TSV file.
}

\section{Endpoint /{apiVersion}/cohorts/create}{

Create a cohort.
}

\section{Endpoint /{apiVersion}/cohorts/search}{

Search cohorts.
}

\section{Endpoint /{apiVersion}/cohorts/{cohorts}/acl}{

Return the acl of the cohort. If member is provided, it will only return the acl for the member.
}

\section{Endpoint /{apiVersion}/cohorts/{cohorts}/delete}{

Delete cohorts.
}

\section{Endpoint /{apiVersion}/cohorts/{cohorts}/info}{

Get cohort information.
}

\section{Endpoint /{apiVersion}/cohorts/{cohorts}/update}{

Update some cohort attributes.
}

\section{Endpoint /{apiVersion}/cohorts/{cohort}/annotationSets/{annotationSet}/annotations/update}{

Update annotations from an annotationSet.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
