% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Family-methods.R
\name{familyClient,OpencgaR-method}
\alias{familyClient,OpencgaR-method}
\title{FamilyClient methods}
\usage{
\S4method{familyClient}{OpencgaR}(
  OpencgaR,
  families,
  family,
  annotationSet,
  members,
  endpointName,
  params = NULL,
  ...
)
}
\arguments{
\item{families}{Comma separated list of family ids.}

\item{family}{Family id.}

\item{annotationSet}{AnnotationSet ID to be updated.}

\item{members}{Comma separated list of individual ids or names.}

\item{creationYear}{Creation year.}

\item{creationMonth}{Creation month (JANUARY, FEBRUARY...).}

\item{creationDay}{Creation day.}

\item{creationDayOfWeek}{Creation day of week (MONDAY, TUESDAY...).}

\item{status}{Status.}

\item{numMembers}{Number of members.}

\item{expectedSize}{Expected size.}

\item{default}{Calculate default stats.}

\item{field}{List of fields separated by semicolons, e.g.: studies;type. For nested fields use >>, e.g.: studies>>biotype;type;numSamples\link{0..10}:1.}

\item{variableSetId}{Variable set ID or name.}

\item{path}{Path where the TSV file is located in OpenCGA or where it should be located.}

\item{parents}{Flag indicating whether to create parent directories if they don't exist (only when TSV file was not previously associated).}

\item{annotationSetId}{Annotation set id. If not provided, variableSetId will be used.}

\item{limit}{Number of results to be returned.}

\item{skip}{Number of results to skip.}

\item{count}{Get the total number of results matching the query. Deactivated by default.}

\item{name}{Family name.}

\item{parentalConsanguinity}{Parental consanguinity.}

\item{samples}{Comma separated list sample IDs or UUIDs up to a maximum of 100.}

\item{phenotypes}{Comma separated list of phenotype ids or names.}

\item{disorders}{Comma separated list of disorder ids or names.}

\item{creationDate}{Creation date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{modificationDate}{Modification date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{annotationsetName}{DEPRECATED: Use annotation queryParam this way: annotationSet[=|==|!|!=]{annotationSetName}.}

\item{variableSet}{DEPRECATED: Use annotation queryParam this way: variableSet[=|==|!|!=]{variableSetId}.}

\item{annotation}{Annotation filters. Example: age>30;gender=FEMALE. For more information, please visit http://docs.opencb.org/display/opencga/AnnotationSets+1.4.0.}

\item{acl}{Filter entries for which a user has the provided permissions. Format: acl={user}:{permissions}. Example: acl=john:WRITE,WRITE_ANNOTATIONS will return all entries for which user john has both WRITE and WRITE_ANNOTATIONS permissions. Only study owners or administrators can query by this field. .}

\item{release}{Release value (Current release from the moment the families were first created).}

\item{snapshot}{Snapshot value (Latest version of families in the specified release).}

\item{member}{User or group id.}

\item{silent}{Boolean to retrieve all possible entries that are queried for, false to raise an exception whenever one of the entries looked for cannot be shown for whichever reason.}

\item{include}{Fields included in the response, whole JSON path must be provided.}

\item{exclude}{Fields excluded in the response, whole JSON path must be provided.}

\item{flattenAnnotations}{Flatten the annotations?.}

\item{version}{Family version.}

\item{deleted}{Boolean to retrieve deleted families.}

\item{updateIndividualVersion}{Update all the individual references from the family to point to their latest versions.}

\item{annotationSetsAction}{Action to be performed if the array of annotationSets is being updated. Allowed values: \link{'ADD', 'SET', 'REMOVE'}}

\item{study}{Study [\link{user@}project:]study where study and project can be either the ID or UUID.}

\item{action}{Action to be performed: ADD to add new annotations; REPLACE to replace the value of an already existing annotation; SET to set the new list of annotations removing any possible old annotations; REMOVE to remove some annotations; RESET to set some annotations to the default value configured in the corresponding variables of the VariableSet if any. Allowed values: \link{'ADD', 'SET', 'REMOVE', 'RESET', 'REPLACE'}}

\item{incVersion}{Create a new version of family.}

\item{updateSampleVersion}{Update all the individual references from the family to point to their latest versions.}

\item{data}{Json containing the map of annotations when the action is ADD, SET or REPLACE, a json with only the key 'remove' containing the comma separated variables to be removed as a value when the action is REMOVE or a json with only the key 'reset' containing the comma separated variables that will be set to the default value when the action is RESET.}
}
\description{
This function implements the OpenCGA calls for managing Families.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   updateAcl \tab /{apiVersion}/families/acl/{members}/update \tab study, members\link{*}, action\link{*}, body\link{*} \cr
   aggregationStats \tab /{apiVersion}/families/aggregationStats \tab study, creationYear, creationMonth, creationDay, creationDayOfWeek, status, phenotypes, release, version, numMembers, expectedSize, annotation, default, field \cr
   loadAnnotationSets \tab /{apiVersion}/families/annotationSets/load \tab study, variableSetId\link{*}, path\link{*}, parents, annotationSetId, body \cr
   create \tab /{apiVersion}/families/create \tab study, members, body\link{*} \cr
   search \tab /{apiVersion}/families/search \tab include, exclude, limit, skip, count, flattenAnnotations, study, name, parentalConsanguinity, members, samples, phenotypes, disorders, creationDate, modificationDate, deleted, annotationsetName, variableSet, annotation, acl, release, snapshot \cr
   acl \tab /{apiVersion}/families/{families}/acl \tab families\link{*}, study, member, silent \cr
   delete \tab /{apiVersion}/families/{families}/delete \tab study, families \cr
   info \tab /{apiVersion}/families/{families}/info \tab include, exclude, flattenAnnotations, families\link{*}, study, version, deleted \cr
   update \tab /{apiVersion}/families/{families}/update \tab families\link{*}, study, incVersion, updateIndividualVersion, annotationSetsAction, body \cr
   updateAnnotations \tab /{apiVersion}/families/{family}/annotationSets/{annotationSet}/annotations/update \tab family\link{*}, study, annotationSet, action, incVersion, updateSampleVersion, body \cr
}
}
\section{Endpoint /{apiVersion}/families/acl/{members}/update}{

Update the set of permissions granted for the member.
}

\section{Endpoint /{apiVersion}/families/aggregationStats}{

Fetch catalog family stats.
}

\section{Endpoint /{apiVersion}/families/annotationSets/load}{

Load annotation sets from a TSV file.
}

\section{Endpoint /{apiVersion}/families/create}{

Create family and the individual objects if they do not exist.
}

\section{Endpoint /{apiVersion}/families/search}{

Search families.
}

\section{Endpoint /{apiVersion}/families/{families}/acl}{

Returns the acl of the families. If member is provided, it will only return the acl for the member.
}

\section{Endpoint /{apiVersion}/families/{families}/delete}{

Delete existing families.
}

\section{Endpoint /{apiVersion}/families/{families}/info}{

Get family information.
}

\section{Endpoint /{apiVersion}/families/{families}/update}{

Update some family attributes.
}

\section{Endpoint /{apiVersion}/families/{family}/annotationSets/{annotationSet}/annotations/update}{

Update annotations from an annotationSet.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
