% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaR-methods.R
\name{initOpencgaR}
\alias{initOpencgaR}
\title{initOpencgaR}
\usage{
initOpencgaR(host = NULL, version = "v2", user = NULL, opencgaConfig = NULL)
}
\arguments{
\item{host}{a character specifying the host url, e.g.
"http://bioinfo.hpc.cam.ac.uk/opencga-prod/"}

\item{version}{a character specifying the API version, e.g. "v2"}
}
\value{
An object of class OpencgaR
}
\description{
This function inicializes the OpencgaR object with the
necessary details to generate a connection.
}
\details{
OpencgaR init function


This class initializes the OpencgaR object. It holds the default configuration
required by OpencgaR methods to connect to OpenCGA web services.
}
\seealso{
\url{https://github.com/opencb/opencga/wiki}
and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\dontrun{
con <- initOpencgaR(host = "http://bioinfo.hpc.cam.ac.uk/opencga-prod/", version = "v2")
con <- opencgaLogin(opencga = con, userid = "demouser", passwd = "demouser", showToken = TRUE)

# Configuration in list format
conf <- list(version="v2", rest=list(host="http://bioinfo.hpc.cam.ac.uk/opencga-prod/"))
con <- initOpencgaR(opencgaConfig=conf)
con <- opencgaLogin(opencga = con, userid = "demouser", passwd = demouser")

# Configuration in file format ("YAML" or "JSON")
conf <- "/path/to/conf/client-configuration.yml"
con <- initOpencgaR(opencgaConfig=conf)
con <- opencgaLogin(opencga = con, userid = "demouser", passwd = "demouser")
}
}
