% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job-methods.R
\name{jobClient,OpencgaR-method}
\alias{jobClient,OpencgaR-method}
\title{JobClient methods}
\usage{
\S4method{jobClient}{OpencgaR}(OpencgaR, job, jobs, members, endpointName, params = NULL, ...)
}
\arguments{
\item{job}{Job ID or UUID.}

\item{jobs}{Comma separated list of job IDs or UUIDs up to a maximum of 100.}

\item{members}{Comma separated list of user or group ids.}

\item{action}{Action to be performed \link{ADD, SET, REMOVE or RESET}.}

\item{toolScope}{Tool scope.}

\item{toolType}{Tool type.}

\item{toolResource}{Tool resource.}

\item{executorId}{Executor id.}

\item{executorFramework}{Executor framework.}

\item{creationYear}{Creation year.}

\item{creationMonth}{Creation month (JANUARY, FEBRUARY...).}

\item{creationDay}{Creation day.}

\item{creationDayOfWeek}{Creation day of week (MONDAY, TUESDAY...).}

\item{status}{Status.}

\item{default}{Calculate default stats.}

\item{field}{List of fields separated by semicolons, e.g.: studies;type. For nested fields use >>, e.g.: studies>>biotype;type;numSamples\link{0..10}:1.}

\item{skip}{Number of results to skip.}

\item{count}{Get the total number of results matching the query. Deactivated by default.}

\item{otherStudies}{Flag indicating the entries being queried can belong to any related study, not just the primary one.}

\item{id}{Job ID. It must be a unique string within the study. An ID will be autogenerated automatically if not provided.}

\item{creationDate}{Creation date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{modificationDate}{Modification date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{visited}{Visited status of job.}

\item{tags}{Job tags.}

\item{input}{Comma separated list of file IDs used as input.}

\item{output}{Comma separated list of file IDs used as output.}

\item{acl}{Filter entries for which a user has the provided permissions. Format: acl={user}:{permissions}. Example: acl=john:WRITE,WRITE_ANNOTATIONS will return all entries for which user john has both WRITE and WRITE_ANNOTATIONS permissions. Only study owners or administrators can query by this field. .}

\item{release}{Release when it was created.}

\item{limit}{Maximum number of jobs to be returned.}

\item{internalStatus}{Job internal status.}

\item{priority}{Priority of the job.}

\item{userId}{User that created the job.}

\item{toolId}{Tool ID executed by the job.}

\item{member}{User or group id.}

\item{silent}{Boolean to retrieve all possible entries that are queried for, false to raise an exception whenever one of the entries looked for cannot be shown for whichever reason.}

\item{include}{Fields included in the response, whole JSON path must be provided.}

\item{exclude}{Fields excluded in the response, whole JSON path must be provided.}

\item{deleted}{Boolean to retrieve deleted jobs.}

\item{data}{body.}

\item{offset}{Starting byte from which the file will be read.}

\item{study}{Study [\link{user@}project:]study where study and project can be either the ID or UUID.}

\item{lines}{Maximum number of lines to be returned.}

\item{type}{Log file to be shown (stdout or stderr).}
}
\description{
This function implements the OpenCGA calls for managing Jobs.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   updateAcl \tab /{apiVersion}/jobs/acl/{members}/update \tab members\link{*}, action\link{*}, body\link{*} \cr
   aggregationStats \tab /{apiVersion}/jobs/aggregationStats \tab study, toolId, toolScope, toolType, toolResource, userId, priority, tags, executorId, executorFramework, creationYear, creationMonth, creationDay, creationDayOfWeek, status, release, default, field \cr
   create \tab /{apiVersion}/jobs/create \tab study, body\link{*} \cr
   search \tab /{apiVersion}/jobs/search \tab include, exclude, limit, skip, count, study, otherStudies, id, toolId, userId, priority, internalStatus, creationDate, modificationDate, visited, tags, input, output, acl, release, deleted \cr
   top \tab /{apiVersion}/jobs/top \tab limit, study, internalStatus, priority, userId, toolId \cr
   acl \tab /{apiVersion}/jobs/{jobs}/acl \tab jobs\link{*}, member, silent \cr
   delete \tab /{apiVersion}/jobs/{jobs}/delete \tab study, jobs \cr
   info \tab /{apiVersion}/jobs/{jobs}/info \tab include, exclude, jobs\link{*}, study, deleted \cr
   update \tab /{apiVersion}/jobs/{jobs}/update \tab jobs\link{*}, study, body \cr
   headLog \tab /{apiVersion}/jobs/{job}/log/head \tab job\link{*}, study, offset, lines, type \cr
   tailLog \tab /{apiVersion}/jobs/{job}/log/tail \tab job\link{*}, study, lines, type \cr
}
}
\section{Endpoint /{apiVersion}/jobs/acl/{members}/update}{

Update the set of permissions granted for the member.
}

\section{Endpoint /{apiVersion}/jobs/aggregationStats}{

Fetch catalog job stats.
}

\section{Endpoint /{apiVersion}/jobs/create}{

Register an executed job with POST method.
}

\section{Endpoint /{apiVersion}/jobs/search}{

Job search method.
}

\section{Endpoint /{apiVersion}/jobs/top}{

Provide a summary of the running jobs.
}

\section{Endpoint /{apiVersion}/jobs/{jobs}/acl}{

Return the acl of the job. If member is provided, it will only return the acl for the member.
}

\section{Endpoint /{apiVersion}/jobs/{jobs}/delete}{

Delete existing jobs.
}

\section{Endpoint /{apiVersion}/jobs/{jobs}/info}{

Get job information.
}

\section{Endpoint /{apiVersion}/jobs/{jobs}/update}{

Update some job attributes.
}

\section{Endpoint /{apiVersion}/jobs/{job}/log/head}{

Show the first lines of a log file (up to a limit).
}

\section{Endpoint /{apiVersion}/jobs/{job}/log/tail}{

Show the last lines of a log file (up to a limit).
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
