% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaR-methods.R
\name{opencgaHelp}
\alias{opencgaHelp}
\alias{OpencgaHelp}
\title{Get help from OpenCGA Web Services}
\usage{
opencgaHelp(opencga, client = NULL, action = NULL)
}
\arguments{
\item{opencga}{an object of type OpencgaR generated using initOpencga or
opencgaLogin}
}
\description{
A function extract all the available information from Opencga web services
at the defined host
}
\examples{
\dontrun{
con <- initOpencgaR(host = "http://bioinfo.hpc.cam.ac.uk/opencga-prod/", version = "v2")
con <- opencgaLogin(opencga = con, userid = "demouser", passwd = "demouser", showToken = TRUE)
opencgaHelp(con, client="userClient", action="info")
}
}
