% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaFunctions.R
\name{OpencgaLogin}
\alias{OpencgaLogin}
\title{A method to query and manipulate user data in Opencga}
\usage{
OpencgaLogin(baseurl, userid = NULL, passwd = NULL, interactive = FALSE)
}
\arguments{
\item{baseurl}{a character specifying the host url}

\item{userid}{a charatcer the username}

\item{passwd}{a charcter the user password}

\item{logical}{whether to launch a graphical interface, FALSE by default}

\item{...}{Any other arguments}
}
\value{
an Opencga class object
}
\description{
This method to query and manipulate user data in Opencga
  
#' @export
OpencgaCreateUser <- function(baseurl, userid, name, passwd, email, organization){
  require(jsonlite)
  userid <- paste0("userid=", userid)
  name <- paste0("name=", name)
  email <- paste0("email=", email)
  passwd <-paste0("password=", passwd)
  organization <- paste0("organization=",organization)
  url <- paste(baseurl, "/users/create?",userid, name, email, organization, passwd)
  res <- fromJSON(url)
  return(res$response$result)
}
A function to login Opencga web services
}

