% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaR-methods.R
\name{opencgaLogin}
\alias{opencgaLogin}
\alias{OpencgaLogin}
\title{Login to OpenCGA Web Services}
\usage{
opencgaLogin(
  opencga,
  userid = NULL,
  passwd = NULL,
  interactive = FALSE,
  autoRenew = FALSE,
  verbose = FALSE,
  showToken = FALSE,
  organization = NULL
)
}
\arguments{
\item{userid}{a charatcer with the username}

\item{passwd}{a charcter with the user password
#@param version a character with the OpenCGA version to use}

\item{interactive}{whether to launch a graphical interface, FALSE by default
#@param ... Any other arguments}

\item{ocga}{an object of type OpencgaR generated using initOpencgaR}
}
\value{
an Opencga class object

\dontrun{
con <- initOpencgaR(host = "http://bioinfo.hpc.cam.ac.uk/opencga-prod/", version = "v2")
con <- opencgaLogin(opencga = con, userid = "xxx", passwd = "xxx", showToken = TRUE)

# Configuration in list format
conf <- list(version="v2", rest=list(host="http://bioinfo.hpc.cam.ac.uk/opencga-prod/"))
con <- initOpencgaR(opencgaConfig=conf)
con <- opencgaLogin(opencga = con, userid = "xxx", passwd = xxx")

# Configuration in file format ("YAML" or "JSON")
conf <- "/path/to/conf/client-configuration.yml"
con <- initOpencgaR(opencgaConfig=conf)
con <- opencgaLogin(opencga = con, userid = "xxx", passwd = "xxx")
}
}
\description{
A function to login Opencga web services
}
