% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaR-methods.R
\name{opencgaLogin}
\alias{opencgaLogin}
\alias{OpencgaLogin}
\title{Login to OpenCGA Web Services}
\usage{
opencgaLogin(opencga, userid = NULL, passwd = NULL, interactive = FALSE)
}
\arguments{
\item{userid}{a charatcer with the username}

\item{passwd}{a charcter with the user password
#@param version a character with the OpenCGA version to use}

\item{interactive}{whether to launch a graphical interface, FALSE by default
#@param ... Any other arguments}

\item{ocga}{an object of type OpencgaR generated using initOpencgaR}
}
\value{
an Opencga class object
}
\description{
A function to login Opencga web services
}
\examples{
con <- initOpencgaR(host = "http://localhost:8080/opencga/", version = "v1", user = "user")
con <- opencgaLogin(opencga = con, userid = "user", passwd = "user_pass")

# Configuration in list format
conf <- list(version="v1", rest=list(host="http://localhost:8080/opencga/"))
con <- initOpencgaR(opencgaConfig=conf)
con <- opencgaLogin(opencga = con, userid = "user", passwd = "user_pass")

# Configuration in file format ("YAML" or "JSON")
conf <- "/path/to/conf/client-configuration.yml"
con <- initOpencgaR(opencgaConfig=conf)
con <- opencgaLogin(opencga = con, userid = "user", passwd = "user_pass")

}
