% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaR-methods.R
\name{opencgaLogin}
\alias{opencgaLogin}
\alias{OpencgaLogin}
\title{Login to OpenCGA Web Services}
\usage{
opencgaLogin(opencga, userid = NULL, passwd = NULL, interactive = FALSE,
  autoRenew = FALSE, showToken = FALSE)
}
\arguments{
\item{userid}{a charatcer with the username}

\item{passwd}{a charcter with the user password
#@param version a character with the OpenCGA version to use}

\item{interactive}{whether to launch a graphical interface, FALSE by default
#@param ... Any other arguments}

\item{ocga}{an object of type OpencgaR generated using initOpencgaR}
}
\description{
A function to login Opencga web services
}
