% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OpencgaR-class}
\alias{OpencgaR-class}
\alias{opencgaR}
\title{OpencgaR Class
#@include commons.R}
\description{
This is an S4 class  which defines the OpencgaR object
}
\details{
This S4 class holds the default configuration required by OpencgaR 
methods to stablish the connection with the web services. By default it is 
configured to query HGVA (http://hgva.opencb.org/).
}
\section{Slots}{

\describe{
\item{\code{host}}{a character specifying the host url. Default 
"http://bioinfo.hpc.cam.ac.uk/hgva"}

\item{\code{version}}{a character specifying the API version. Default "v1"}
}}

\seealso{
\url{https://github.com/opencb/opencga/wiki} 
and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
