% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OpencgaR-class}
\alias{OpencgaR-class}
\alias{opencgaR}
\title{OpencgaR Class}
\description{
This is an S4 class which defines the OpencgaR object
}
\details{
This S4 class holds the default configuration required by OpencgaR 
methods to establish the connection with the web services.
}
\section{Slots}{

\describe{
\item{\code{host}}{a character specifying the host URL. Example: 
"http://bioinfo.hpc.cam.ac.uk/opencga-prod/"}

\item{\code{version}}{a character specifying the API version. Example: "v2"}

\item{\code{user}}{a character string with you OpenCGA username}

\item{\code{token}}{string containing your token}

\item{\code{refreshToken}}{string containing your token}

\item{\code{autoRenew}}{Boolean specifying if the token should be renewed 
automatically if the session is going to expire in less than 5 minutes}

\item{\code{verbose}}{Boolean specifying if logs should be displayed. Logs include: 
URL (with or without token - see param showToken), query status and number 
of documents retrieved}

\item{\code{showToken}}{Boolean specifying if the token should be displayed in the 
verbose mode}

\item{\code{swagger}}{list with the swagger information of the OpenCGA instance in the
specified host}
}}

\seealso{
\url{https://github.com/opencb/opencga/wiki} 
and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
}
