% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OpencgaR-class}
\alias{OpencgaR-class}
\alias{opencgaR}
\title{OpencgaR Class}
\description{
This is an S4 class which defines the OpencgaR object
}
\details{
This S4 class holds the default configuration required by OpencgaR 
methods to stablish the connection with the web services.
}
\section{Slots}{

\describe{
\item{\code{host}}{a character specifying the host url. Example: 
"http://bioinfo.hpc.cam.ac.uk/opencga-prod/"}

\item{\code{version}}{a character specifying the API version. Example: "v2"}

\item{\code{user}}{a character string with you OpenCGA username}

\item{\code{token}}{string containin your token}

\item{\code{refreshToken}}{string containin your renew token}

\item{\code{swagger}}{list with the swagger information of the OpenCGA intance in the
spcified host}
}}

\seealso{
\url{https://github.com/opencb/opencga/wiki} 
and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
}
