% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaR-methods.R
\name{opencgaReadConfig}
\alias{opencgaReadConfig}
\title{Read OpenCGA configuration}
\usage{
opencgaReadConfig(conf)
}
\arguments{
\item{conf}{a list or the path to a file (in "JSON" or "YAML" format)
containing the host and version configurations to set OpenCGA connection}
}
\value{
a OpencgaR object

\dontrun{
con <- initOpencgaR(host = "http://bioinfo.hpc.cam.ac.uk/opencga-prod/", version = "v2")
con <- opencgaLogin(opencga = con, userid = "demouser", passwd = "demouser", showToken = TRUE)

# Configuration in list format
conf <- list(version="v2", rest=list(host="http://bioinfo.hpc.cam.ac.uk/opencga-prod/"))
con <- initOpencgaR(opencgaConfig=conf)
con <- opencgaLogin(opencga = con, userid = "demouser", passwd = demouser")

# Configuration in file format ("YAML" or "JSON")
conf <- "/path/to/conf/client-configuration.yml"
con <- initOpencgaR(opencgaConfig=conf)
con <- opencgaLogin(opencga = con, userid = "demouser", passwd = "demouser")
}
}
\description{
Read OpenCGA configuration
}
