% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project-methods.R
\docType{methods}
\name{projectClient,OpencgaR-method}
\alias{projectClient,OpencgaR-method}
\title{ProjectClient methods}
\usage{
\S4method{projectClient}{OpencgaR}(OpencgaR, project, action, params = NULL)
}
\description{
This function implements the OpenCGA calls for managing Projects
}
\section{Slots}{

\describe{
\item{\code{OpencgaR}}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{\code{project}}{a character or a vector contaning the project ids or aliases to 
work on. Mandatory when using "info", "studies", "delete", "increlease" and 
"update" actions}

\item{\code{action}}{action to be performed on the projects}

\item{\code{params}}{list containing additional query or body params}
}}

\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
