% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project-methods.R
\name{projectClient,OpencgaR-method}
\alias{projectClient,OpencgaR-method}
\title{ProjectClient methods}
\usage{
\S4method{projectClient}{OpencgaR}(OpencgaR, projects, project, endpointName, params = NULL, ...)
}
\arguments{
\item{projects}{Comma separated list of projects \link{user@}project up to a maximum of 100.}

\item{project}{Project \link{user@}project where project can be either the ID or the alias.}

\item{owner}{Owner of the project.}

\item{id}{Project \link{user@}project where project can be either the ID or the alias.}

\item{name}{Project name.}

\item{fqn}{Project fqn.}

\item{organization}{Project organization.}

\item{description}{Project description.}

\item{study}{Study id.}

\item{creationDate}{Creation date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{modificationDate}{Modification date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{status}{Status.}

\item{attributes}{Attributes.}

\item{default}{Calculate default stats.}

\item{fileFields}{List of file fields separated by semicolons, e.g.: studies;type. For nested fields use >>, e.g.: studies>>biotype;type.}

\item{individualFields}{List of individual fields separated by semicolons, e.g.: studies;type. For nested fields use >>, e.g.: studies>>biotype;type.}

\item{familyFields}{List of family fields separated by semicolons, e.g.: studies;type. For nested fields use >>, e.g.: studies>>biotype;type.}

\item{sampleFields}{List of sample fields separated by semicolons, e.g.: studies;type. For nested fields use >>, e.g.: studies>>biotype;type.}

\item{cohortFields}{List of cohort fields separated by semicolons, e.g.: studies;type. For nested fields use >>, e.g.: studies>>biotype;type.}

\item{jobFields}{List of job fields separated by semicolons, e.g.: studies;type. For nested fields use >>, e.g.: studies>>biotype;type.}

\item{include}{Fields included in the response, whole JSON path must be provided.}

\item{exclude}{Fields excluded in the response, whole JSON path must be provided.}

\item{limit}{Number of results to be returned.}

\item{skip}{Number of results to skip.}

\item{data}{JSON containing the params to be updated. It will be only possible to update organism fields not previously defined.}
}
\description{
This function implements the OpenCGA calls for managing Projects.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   create \tab /{apiVersion}/projects/create \tab body\link{*} \cr
   search \tab /{apiVersion}/projects/search \tab include, exclude, limit, skip, owner, id, name, fqn, organization, description, study, creationDate, modificationDate, status, attributes \cr
   aggregationStats \tab /{apiVersion}/projects/{projects}/aggregationStats \tab projects\link{*}, default, fileFields, individualFields, familyFields, sampleFields, cohortFields, jobFields \cr
   info \tab /{apiVersion}/projects/{projects}/info \tab include, exclude, projects\link{*} \cr
   incRelease \tab /{apiVersion}/projects/{project}/incRelease \tab project\link{*} \cr
   studies \tab /{apiVersion}/projects/{project}/studies \tab include, exclude, limit, skip, project\link{*} \cr
   update \tab /{apiVersion}/projects/{project}/update \tab project\link{*}, body\link{*} \cr
}
}
\section{Endpoint /{apiVersion}/projects/create}{

Create a new project.
}

\section{Endpoint /{apiVersion}/projects/search}{

Search projects.
}

\section{Endpoint /{apiVersion}/projects/{projects}/aggregationStats}{

Fetch catalog project stats.
}

\section{Endpoint /{apiVersion}/projects/{projects}/info}{

Fetch project information.
}

\section{Endpoint /{apiVersion}/projects/{project}/incRelease}{

Increment current release number in the project.
}

\section{Endpoint /{apiVersion}/projects/{project}/studies}{

Fetch all the studies contained in the project.
}

\section{Endpoint /{apiVersion}/projects/{project}/update}{

Update some project attributes.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
