% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-methods.R
\name{sampleClient,OpencgaR-method}
\alias{sampleClient,OpencgaR-method}
\title{SampleClient methods}
\usage{
\S4method{sampleClient}{OpencgaR}(
  OpencgaR,
  sample,
  annotationSet,
  members,
  samples,
  endpointName,
  params = NULL,
  ...
)
}
\arguments{
\item{sample}{Sample ID.}

\item{annotationSet}{AnnotationSet ID to be updated.}

\item{members}{Comma separated list of user or group ids.}

\item{samples}{Comma separated list sample IDs or UUIDs up to a maximum of 100.}

\item{propagate}{Propagate sample permissions to related individuals.}

\item{source}{Source.}

\item{creationYear}{Creation year.}

\item{creationMonth}{Creation month (JANUARY, FEBRUARY...).}

\item{creationDay}{Creation day.}

\item{creationDayOfWeek}{Creation day of week (MONDAY, TUESDAY...).}

\item{status}{Status.}

\item{type}{Type.}

\item{default}{Calculate default stats.}

\item{field}{List of fields separated by semicolons, e.g.: studies;type. For nested fields use >>, e.g.: studies>>biotype;type;numSamples\link{0..10}:1.}

\item{variableSetId}{Variable set ID or name.}

\item{path}{Path where the TSV file is located in OpenCGA or where it should be located.}

\item{parents}{Flag indicating whether to create parent directories if they don't exist (only when TSV file was not previously associated).}

\item{annotationSetId}{Annotation set id. If not provided, variableSetId will be used.}

\item{individual}{DEPRECATED: It should be passed in the body.}

\item{file}{file.}

\item{variableSet}{variableSet.}

\item{limit}{Number of results to be returned.}

\item{skip}{Number of results to skip.}

\item{count}{Get the total number of results matching the query. Deactivated by default.}

\item{id}{Comma separated list sample IDs or UUIDs up to a maximum of 100.}

\item{somatic}{Somatic sample.}

\item{individualId}{Individual ID.}

\item{creationDate}{Creation date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{modificationDate}{Modification date. Format: yyyyMMddHHmmss. Examples: >2018, 2017-2018, <201805.}

\item{phenotypes}{Comma separated list of phenotype ids or names.}

\item{annotation}{Annotation filters. Example: age>30;gender=FEMALE. For more information, please visit http://docs.opencb.org/display/opencga/AnnotationSets+1.4.0.}

\item{acl}{Filter entries for which a user has the provided permissions. Format: acl={user}:{permissions}. Example: acl=john:WRITE,WRITE_ANNOTATIONS will return all entries for which user john has both WRITE and WRITE_ANNOTATIONS permissions. Only study owners or administrators can query by this field. .}

\item{attributes}{Text attributes (Format: sex=male,age>20 ...).}

\item{release}{Release when it was created.}

\item{snapshot}{Snapshot value (Latest version of the entry in the specified release).}

\item{member}{User or group id.}

\item{silent}{Boolean to retrieve all possible entries that are queried for, false to raise an exception whenever one of the entries looked for cannot be shown for whichever reason.}

\item{force}{Force the deletion of samples even if they are associated to files, individuals or cohorts.}

\item{emptyFilesAction}{Action to be performed over files that were associated only to the sample to be deleted. Possible actions are NONE, TRASH, DELETE.}

\item{deleteEmptyCohorts}{Boolean indicating if the cohorts associated only to the sample to be deleted should be also deleted.}

\item{include}{Fields included in the response, whole JSON path must be provided.}

\item{exclude}{Fields excluded in the response, whole JSON path must be provided.}

\item{includeIndividual}{Include Individual object as an attribute.}

\item{flattenAnnotations}{Flatten the annotations?.}

\item{version}{Sample version.}

\item{deleted}{Boolean to retrieve deleted entries.}

\item{annotationSetsAction}{Action to be performed if the array of annotationSets is being updated. Allowed values: \link{'ADD', 'SET', 'REMOVE'}}

\item{study}{Study [\link{user@}project:]study where study and project can be either the ID or UUID.}

\item{action}{Action to be performed: ADD to add new annotations; REPLACE to replace the value of an already existing annotation; SET to set the new list of annotations removing any possible old annotations; REMOVE to remove some annotations; RESET to set some annotations to the default value configured in the corresponding variables of the VariableSet if any. Allowed values: \link{'ADD', 'SET', 'REMOVE', 'RESET', 'REPLACE'}}

\item{incVersion}{Create a new version of sample.}

\item{data}{Json containing the map of annotations when the action is ADD, SET or REPLACE, a json with only the key 'remove' containing the comma separated variables to be removed as a value when the action is REMOVE or a json with only the key 'reset' containing the comma separated variables that will be set to the default value when the action is RESET.}
}
\description{
This function implements the OpenCGA calls for managing Samples.
The following table summarises the available \emph{actions} for this client:\tabular{llr}{
   endpointName \tab Endpoint WS \tab parameters accepted \cr
   updateAcl \tab /{apiVersion}/samples/acl/{members}/update \tab study, members\link{*}, action\link{*}, propagate, body\link{*} \cr
   aggregationStats \tab /{apiVersion}/samples/aggregationStats \tab study, source, creationYear, creationMonth, creationDay, creationDayOfWeek, status, type, phenotypes, release, version, somatic, annotation, default, field \cr
   loadAnnotationSets \tab /{apiVersion}/samples/annotationSets/load \tab study, variableSetId\link{*}, path\link{*}, parents, annotationSetId, body \cr
   create \tab /{apiVersion}/samples/create \tab study, individual, body\link{*} \cr
   load \tab /{apiVersion}/samples/load \tab study, file\link{*}, variableSet \cr
   search \tab /{apiVersion}/samples/search \tab include, exclude, limit, skip, count, includeIndividual, flattenAnnotations, study, id, somatic, individualId, creationDate, modificationDate, phenotypes, annotation, acl, attributes, release, deleted, snapshot \cr
   acl \tab /{apiVersion}/samples/{samples}/acl \tab samples\link{*}, study, member, silent \cr
   delete \tab /{apiVersion}/samples/{samples}/delete \tab force, emptyFilesAction, deleteEmptyCohorts, study, samples \cr
   info \tab /{apiVersion}/samples/{samples}/info \tab include, exclude, includeIndividual, flattenAnnotations, samples\link{*}, study, version, deleted \cr
   update \tab /{apiVersion}/samples/{samples}/update \tab samples\link{*}, study, incVersion, annotationSetsAction, body \cr
   updateAnnotations \tab /{apiVersion}/samples/{sample}/annotationSets/{annotationSet}/annotations/update \tab sample\link{*}, study, annotationSet, action, incVersion, body \cr
}
}
\section{Endpoint /{apiVersion}/samples/acl/{members}/update}{

Update the set of permissions granted for the member.
}

\section{Endpoint /{apiVersion}/samples/aggregationStats}{

Fetch catalog sample stats.
}

\section{Endpoint /{apiVersion}/samples/annotationSets/load}{

Load annotation sets from a TSV file.
}

\section{Endpoint /{apiVersion}/samples/create}{

Create sample.
}

\section{Endpoint /{apiVersion}/samples/load}{

Load samples from a ped file \link{EXPERIMENTAL}.
}

\section{Endpoint /{apiVersion}/samples/search}{

Sample search method.
}

\section{Endpoint /{apiVersion}/samples/{samples}/acl}{

Returns the acl of the samples. If member is provided, it will only return the acl for the member.
}

\section{Endpoint /{apiVersion}/samples/{samples}/delete}{

Delete samples.
}

\section{Endpoint /{apiVersion}/samples/{samples}/info}{

Get sample information.
}

\section{Endpoint /{apiVersion}/samples/{samples}/update}{

Update some sample attributes.
}

\section{Endpoint /{apiVersion}/samples/{sample}/annotationSets/{annotationSet}/annotations/update}{

Update annotations from an annotationSet.
}

\seealso{
\url{http://docs.opencb.org/display/opencga/Using+OpenCGA} and the RESTful API documentation
\url{http://bioinfo.hpc.cam.ac.uk/opencga-prod/webservices/}
\link{*}: Required parameter
}
