% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study-methods.R
\docType{methods}
\name{studyClient,OpencgaR-method}
\alias{studyClient,OpencgaR-method}
\title{StudyClient methods}
\usage{
\S4method{studyClient}{OpencgaR}(OpencgaR, study, action, params = NULL)
}
\description{
This function implements the OpenCGA calls for managing Studies
}
\section{Slots}{

\describe{
\item{\code{OpencgaR}}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{\code{study}}{a character string of the study to work on, in the format 
[[user@]project:]study where study and project can be either the id or alias. 
Mandatory when using studyClient with all actions except "create" and "search" 
and all actions in studyGroupClient}

\item{\code{group}}{a character string containing the groupId to work with. Mandatory 
when using studyGroupClient "delete" action. If specified when using "update", 
updates the members of the group, otherwise, is used to add/remove users with 
access to study}

\item{\code{memberId}}{a character or a vector contaning user or group ids to 
work on. Mandatory when using studyAclClient}

\item{\code{action}}{action to be performed on the studies}

\item{\code{params}}{list containing additional query or body params}
}}

\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
