% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User-methods.R
\docType{methods}
\name{userClient,OpencgaR-method}
\alias{userClient,OpencgaR-method}
\title{UserClient methods}
\usage{
\S4method{userClient}{OpencgaR}(OpencgaR, user, action, params = NULL)
}
\description{
This function implements the OpenCGA calls for managing Users
}
\section{Slots}{

\describe{
\item{\code{OpencgaR}}{an object OpencgaR generated using initOpencgaR and/or opencgaLogin 
where the connection and session details are stored}

\item{\code{user}}{a character or a vector contaning the user ids or aliases to 
work on. Mandatory in all actions except "create"}

\item{\code{name}}{a character or a vector contaning the filter name to work on. 
Mandatory in all userConfigClient actions except "create" and all 
userFilterClient except "list" and "create"}

\item{\code{action}}{action to be performed on users}

\item{\code{params}}{list containing additional query or body params}
}}

\seealso{
\url{https://github.com/opencb/opencga/wiki} and the RESTful API documentation 
\url{http://bioinfo.hpc.cam.ac.uk/opencga/webservices/}
}
